% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{alpha_max_loo}
\alias{alpha_max_loo}
\title{Maximize leave-one-out with respect to alpha}
\usage{
alpha_max_loo(counts, alpha0, clust_mem, weights, psc = 1e-10,
  eps = 1e-04, max_iter = 10000, threads = 1)
}
\arguments{
\item{counts}{A gene by droplet sparse count matrix.}

\item{alpha0}{A gene by cluster matrix of initial alpha parameter values.}

\item{clust_mem}{A droplet by cluster boolean matrix, indicating whether 
a droplet will count towards estimation of the cluster's parameters.}

\item{weights}{A droplet by cluster matrix of weights, or the posterior 
probability the droplet belongs to the cluster}

\item{psc}{pseudocount to add to the final alpha. This avoids alpha values 
of 0 that are outside the domain of the parameter.}

\item{eps}{The threshold of change in the parameter to stop the 
LOO iterations.}

\item{max_iter}{The maximum number of LOO iterations.}

\item{threads}{Number of threads for parallel execution. Default is 1.}
}
\value{
A gene by cluster matrix of alpha values for the cluster's 
 DM parameters
}
\description{
Estimate Alpha parameters of DM using method of moments. The input 
data \code{counts} is in sparse matrix format. The parameters for 
DM can be estimated for multiple centers, where both \code{clust_mem} and 
\code{weights} specify the cluster's members and their weights, 
respectively.
}
