% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{alpha_mom}
\alias{alpha_mom}
\title{Estimate alpha using method of moments}
\usage{
alpha_mom(counts, clust_mem, weights, psc = 1e-10, threads = 1)
}
\arguments{
\item{counts}{A gene by droplet sparse count matrix.}

\item{clust_mem}{A droplet by cluster boolean matrix, indicating whether 
a droplet will count towards estimation of the cluster's parameters.}

\item{weights}{A droplet by cluster matrix of weights, or the posterior 
probability the droplet belongs to the cluster}

\item{psc}{pseudocount to add to the final alpha. This avoids alpha values 
of 0 that are outside the domain of the parameter.}

\item{threads}{Number of threads for parallel execution. Default is 1.}
}
\value{
A gene by cluster matrix of alpha values for the cluster's 
 DM parameters
}
\description{
Estimate Alpha parameters of DM using method of moments. The input 
data \code{counts} is in sparse matrix format. The parameters for 
DM can be estimated for multiple centers, where both \code{clust_mem} and 
\code{weights} specify the cluster's members and their weights, 
respectively.
}
