% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{assign_clusters}
\alias{assign_clusters}
\title{Assign clusters to droplets from EM}
\usage{
assign_clusters(x, k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{k_init}{The k_init run to call droplets from. This must 
specify a single value. If left unspecified, will call droplets only if 
there is one k_init run.}
}
\value{
An SCE object.
}
\description{
Assign cluster to droplets in the test_data from a DIEM k_init run.
}
