% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_debris_clusters}
\alias{call_debris_clusters}
\title{Call debris clusters using the debris score}
\usage{
call_debris_clusters(x, thresh = 0.25, top_n = 20, deb_clust = NULL,
  k_init = NULL, verbose = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{thresh}{Clusters with a fraction of debris droplets above this 
value are set to debris clusters.}

\item{top_n}{Number of differentially expressed genes per cluster 
to use for calculating debris score.}

\item{deb_clust}{Manually specify clusters which should be set as debris 
clusters.}

\item{k_init}{The k_init run to call droplets from. This must 
specify a single value. If left unspecified, will call droplets only if 
there is one k_init run.}

\item{verbose}{verbosity}
}
\value{
An SCE object.
}
\description{
Specify which clusters from DIEM are debris clusters. This uses the 
debris score of droplets from \code{\link{estimate_db_score}}. The 
average debris score of clusters assigned to the fixed debris cluster 
(cluster 1) is calculated. Then, for each cluster, the fraction of 
droplets with a debris score above this mean value is calculated. 
In addition to the fixed debris cluster, any clusters with a 
fraction of debris droplets above \code{thresh} (0.5 by default) 
are assigned as debris clusters.
}
