% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call clean droplets after running EM}
\usage{
call_targets(x, pp_thresh = 0.95, min_genes = 200)
}
\arguments{
\item{x}{An SCE object.}

\item{pp_thresh}{Numeric threshold, where clean droplets must have a 
posterior probability of at least \code{pp_thresh}.}

\item{min_genes}{Numeric threshold, where clean droplets must have at least 
\code{min_genes} genes detected.}
}
\value{
An SCE object.
}
\description{
Call targets from droplets if the log-likelihood membership 
probability (the posterior probability) is higher than \code{pp_thresh}.
}
