% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_k_init}
\alias{check_k_init}
\title{Check values of given k_init against kruns}
\usage{
check_k_init(x, k_init, return_all = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{k_init}{Run EM on the \code{k_init} initialization(s). 
If NULL (default), run on all \code{k_init} initializations.}

\item{return_all}{If there are more than one k_init values, return 
all or throw an error.}
}
\value{
integer specifying the k_init value(s).
}
\description{
Check values of given k_init against kruns
}
