% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{clusters}
\alias{clusters}
\title{Get cluster labels}
\usage{
clusters(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
vector of cluster labels, NULL if no clusters
}
\description{
Get cluster labels
}
