% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{condense_labs}
\alias{condense_labs}
\title{Re-assign labels by condensing to 1 to k}
\usage{
condense_labs(labs)
}
\arguments{
\item{labs}{Vector of labels.}
}
\value{
Vector of labels in which the numeric values are contiguous.
}
\description{
Re-assign labels by condensing to 1 to k
}
