% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{contour_plot_call}
\alias{contour_plot_call}
\title{Contour plot of PCs + bg_score, grouped by call}
\usage{
contour_plot_call(sce, n_pcs = 4, return = FALSE)
}
\arguments{
\item{return}{Boolean. Return a ggplot object}

\item{x}{SCE. SCE object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Contour plot of PCs + bg_score, grouped by call
}
