% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{contour_plot_pi}
\alias{contour_plot_pi}
\title{Contour plot of pi values}
\usage{
contour_plot_pi(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{return}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Contour plot of pi values
}
