% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{contour_plot_sim}
\alias{contour_plot_sim}
\title{Contour plot of PCs + bg_score, grouped by sim}
\usage{
contour_plot_sim(sce, return = FALSE)
}
\arguments{
\item{return}{Boolean. Return a ggplot object}

\item{x}{SCE. SCE object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Contour plot of PCs + bg_score, grouped by sim
}
