% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{convert_to_seurat}
\alias{convert_to_seurat}
\title{Convert SCE object to Seurat}
\usage{
convert_to_seurat(x, targets = TRUE, meta = TRUE, ...)
}
\arguments{
\item{x}{SCE object}

\item{targets}{Boolean. If \code{TRUE}, then output droplets that are specified as targets}

\item{meta}{Boolean. Place information in dropl_info slot into meta.data in Seurat object}

\item{...}{Arguments to \code{\link[Seurat]{CreateSeuratObject}}, such as \code{project} for project name}
}
\value{
A Seurat object
}
\description{
Convert SCE object to Seurat. if \code{targets} is true (default), outputs only droplets that are 
called targets. In addition, outputs meta information calculated, such as bg_score. Additional arguments to 
this function are passed onto \code{\link[Seurat]{CreateSeuratObject}}. Common arguments include 
"min.cells = 3" and "min.features = 200".
}
