% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{correct_counts}
\alias{correct_counts}
\title{Correct counts by subtracting estimated counts from debris}
\usage{
correct_counts(x, deb_clust = c(1), round_count = TRUE,
  k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{deb_clust}{A numeric vector of debris clusters. Contains 
only cluster 1 by default, and should always contain cluster 1.}

\item{round_count}{Boolean indiciating whether to round the counts 
to integers or not.}

\item{k_init}{Run EM on the \code{k_init} initialization(s). 
If NULL (default), run on all \code{k_init} initializations.}
}
\description{
Remove the counts predicted to originate from debris clusters. This 
function takes the cluster proportions for each droplet, and estimated 
the debris counts by adding the expresion means for the debris clusters 
weighted by the their proportions mentioned above, and subtracting this
value from the counts. The debris clusters are specified by deb_clust 
and default to cluster 1, which should always be included. The values to 
subtract are rounded according to the function give by the parameter rfunc, 
which is round by default. This function can be floor, so that only debris 
counts greater than or equal to 1 are subtracted. The rfunc function 
can also be set to NULL, so that no rounding is performed.
}
