% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{ddm_sparse}
\alias{ddm_sparse}
\title{Log density of Dirichlet-multinomial
X is a sample by gene matrix}
\usage{
ddm_sparse(X, alpha)
}
\description{
Log density of Dirichlet-multinomial
X is a sample by gene matrix
}
