% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{ddm_sparseIM}
\alias{ddm_sparseIM}
\title{Log density of Dirichlet-multinomial
X is a sample by gene matrix}
\usage{
ddm_sparseIM(X, alpha)
}
\description{
Log density of Dirichlet-multinomial
X is a sample by gene matrix
}
