% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{de_ttest}
\alias{de_ttest}
\title{Run a DE between two groups with Welch's t-test}
\usage{
de_ttest(x, c1, c2)
}
\arguments{
\item{x}{A sparse matrix}

\item{c1}{Column indexes or names of first group}

\item{c2}{Column indexes or names of second group}
}
\value{
A data frame of DE results with
\describe{
\item{diff}{difference between mu1 and mu2}
\item{logFC}{log2 fold change}
\item{mu1}{mean of group 1}
\item{mu2}{mean of group 2}
\item{t}{t statistic}
\item{p}{p-value}
\item{gene}{gene name}
}
}
\description{
Run DE between two groups. \code{x} is a sparse matrix with 
normalized expression data (typically proportions). \code{c1} 
and \code{c2} are vectors containing indices or names 
of the columns that belong to group 1 and 2, respectively. 
A Welch's t-test is performed.
}
