% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{de_ttest_all}
\alias{de_ttest_all}
\title{Run a DE across clusters with Welch's t-test}
\usage{
de_ttest_all(counts, labels, normalize = TRUE, sf = 1,
  p_thresh = 0.05, logfc_thresh = 0, p_method = "bonferroni")
}
\arguments{
\item{counts}{Gene by droplet sparse matrix of expression data.}

\item{labels}{Vector of droplet labels giving the cluster that the 
droplet belongs to.}

\item{normalize}{Whether to scale the droplets to sum to \code{sf} and 
and then log transform.}

\item{sf}{Scale the counts in each droplet to sum to this value.}

\item{p_thresh}{Return only genes that have an adjusted p-value less than 
this value.}

\item{logfc_thresh}{Return genes that have a log fold change of at least 
this value.}

\item{p_method}{Adjust p-values for multiple testing using this method.}
}
\value{
A data frame of DE results with
\describe{
\item{gene}{gene name}
\item{diff}{difference between mu1 and mu2}
\item{logFC}{log2 fold change}
\item{mu1}{mean of group 1}
\item{mu2}{mean of group 2}
\item{t}{t statistic}
\item{p}{p-value}
\item{p_adj}{p-value corrected for multiple testing}
\item{cluster}{cluster the DE genes belong to}
}
}
\description{
Given expression data in \code{counts}, run differential 
expression between droplets in one group against all others, for 
each group in \code{labels}.
These counts can be normalized, or raw counts 
can be given. If providiing raw counts, set the \code{normalize} 
parameter to TRUE. Returns only genes with an adjust p-value 
less than \code{p_thresh} and a log fold change greater than 
\code{logFC}.
}
\examples{
\donttest{
counts <- raw_counts(sce, type = "test")
clusts <- clusters(sce)
markers <- de_ttest_all(counts, clusts)
}
}
