% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{debris_genes}
\alias{debris_genes}
\title{Return the differentially expression results for debris genes}
\usage{
debris_genes(x, p_adj = 0.05, logFC = 0)
}
\arguments{
\item{x}{An SCE object.}

\item{p_adj}{Return only rows (genes) with an adjusted p-value less 
than this value.}

\item{logFC}{Return only rows (genes) with a log fold-change greater 
than this value.}
}
\value{
A data frame
}
\description{
Return the data frame stored in the slot \code{debris_genes}. This 
contains the results of the differential expression analysis, 
including
\describe{
\item{diff}{Difference in proportion of gene expression between 
 the droplets in the debris and cell type clusters}
\item{logFC}{log2 fold change between the droplets in the debris 
 and cell type clusters}
\item{mu1}{Mean of gene expression proportion in the debris droplets}
\item{mu2}{Mean of gene expression proportion in the cell type droplets}
\item{t}{t statistic}
\item{p}{P value}
\item{gene}{Gene name}
\item{p_adj}{P value adjusted for multiple testing}
}
}
