% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{DIEM_10X}
\alias{DIEM_10X}
\title{Run DIEM pipeline directly from 10X output}
\usage{
DIEM_10X(path_10X, min_bg_count = 25, max_bg_count = 150, n_pcs = 5,
  expected_targets = 10000, min_count = 200, p = 0.5, seedn = NULL)
}
\description{
Run DIEM pipeline directly from 10X output
}
