% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{diem_init}
\alias{diem_init}
\title{Initialize DIEM}
\usage{
diem_init(sce, top_n = 10000, min_counts = 100,
  init_de_log_base = 50, init_de_cutpoint = NULL, cpm_thresh = 3,
  log2fc_thresh = 0, init_em_tails = 0.05, lk_fraction = 0.95,
  ss_pct = 0, verbose = TRUE)
}
\arguments{
\item{sce}{SCE. SCE object with raw counts.}

\item{top_n}{Numeric. Only the top \code{top_n} droplets ranked by total counts are classified.}

\item{min_counts}{Integer. Only droplets with at least this number of total counts are classified. 
Can override \code{top_n} if \code{top_n} includes droplets with less than \code{min_counts}.}

\item{init_de_log_base}{Numeric. Take droplets with counts within 2 logs below the max droplet size using this log base
for initializing the DE cutpoint.}

\item{init_de_cutpoint}{Numeric. Manually set the initialization of the DE cutpoint. If not NULL, it is calculated from the data.}

\item{cpm_thresh}{Numeric. Calculate pi from genes with counts per million mapped reads (CPM) above this threshold.}

\item{log2fc_thresh}{Numeric. Calculate pi from genes with log2 fold changes above this threshold.}

\item{lk_fraction}{Numeric. Select droplets that have a fraction of target likelihood greater than this number.}

\item{verbose}{Boolean. Verbosity.}
}
\value{
SCE object.
}
\description{
Initialize DIEM pipeline on an SCE object, e.g. one returned from \code{\link{read_10x}} function.
}
