% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{distances}
\alias{distances}
\title{Get cluster distances to background}
\usage{
distances(x, k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{k_init}{The k_init run to extract the values from.}
}
\value{
A numeric vector containing the distances for each cluster.
}
\description{
Get cluster distances to background
}
