% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{divide_by_colsum}
\alias{divide_by_colsum}
\title{Divide elements of a column in a sparse matrix by its column sum}
\usage{
divide_by_colsum(x)
}
\arguments{
\item{x}{Sparse Matrix}
}
\value{
A Sparse Matrix with columns that sum to 1.
}
\description{
Divide elements of a column in a sparse matrix by its column sum
}
