% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{dm_loo}
\alias{dm_loo}
\title{LOO optimization}
\usage{
dm_loo(counts, A, Z, eps = 1e-04, max_loo = 3, add_to = 1e-16,
  verbose = FALSE)
}
\description{
counts is a droplet by gene matrix
sizes is total droplet size
a is current value of alpha parameter, gene by k matrix
z is current membership
}
