% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{dm_nr}
\alias{dm_nr}
\title{Newton-Raphson optimization for Dir-Mult with sparse matrix}
\usage{
dm_nr(x, a, z)
}
\description{
x is gene by droplet matrix
a is current value of alpha parameter, gene by k matrix
z is current membership
}
