% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{dmmn_llk}
\alias{dmmn_llk}
\title{Computed log likelihood of multinomial mixture for a matrix}
\usage{
dmmn_llk(x, p, tau, probs = NULL, labels = NULL)
}
\arguments{
\item{x}{sparseMatrix. A sample by feature matrix of counts.}

\item{p}{numeric matrix. A feature by group matrix of probabilities.}

\item{log}{Logical. If TRUE, return matrix with log probabilities.}
}
\value{
a numeric matrix with n samples by x variables.
}
\description{
Given an n x p matrix of count data \code{x}, as well as a 
p x k matrix of probabilities \code{p}, return the probabilities 
of each sample for the k groups in a n x k matrix. If \code{x} 
has any non-integer values, they are rounded to the nearest integer.
}
