% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{dmultinom_sparse}
\alias{dmultinom_sparse}
\title{Get multinomial log prob of columns in X}
\usage{
dmultinom_sparse(X, prob)
}
\arguments{
\item{X}{sparseMatrix. A sample by feature matrix of counts.}

\item{prob}{Numeric}
}
\description{
Get multinomial log prob of columns in X
}
