% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{draw_diri_multinom}
\alias{draw_diri_multinom}
\title{Create matrix with dirichlet-multinomial draws}
\usage{
draw_diri_multinom(bins, bin_sampled, size_sampled, n_draw, alpha = 100,
  seedn = NULL)
}
\arguments{
\item{bins}{Matrix. Contains probabilities for random samples for dirichlet}

\item{bin_sampled}{Vector. Column indices of bins to use for each draw}

\item{size_sampled}{Vector. Sizes to use for each multinomial draw}

\item{n_draw}{Integer. Number of random samples to take}

\item{alpha}{Numeric. Value to multiply the probability vector in bins to get dirichlet parameter.}

\item{seedn}{Boolean. Seed number}
}
\value{
Matrix with each column containing a random sample from the dirichlet-multinomial.
}
\description{
Draws random samples from a dirichlet-multinomial distribution. The \code{bins} 
matrix contains the probability values for sampling in the columns. The 
columns to use for each draw are given in the vector \code{bin_sampled}. This should 
be the same length as the value of \code{n_draw}. The dirichlet is sampled by 
taking a sample from the gamma distribution, using the probability in \code{bins} 
as the shape. The parameter \code{alpha} multiplies the probability vector in 
\code{bins}. The larger the value, the tighter the dirichlet sample, leading to 
the dirichlet-multinomial approaching the multinomial. After sampling the dirichlet, 
a multinomial of size taken from the vector \code{size_sampled} is taken.
}
