% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{draw_multinom}
\alias{draw_multinom}
\title{Create matrix with multinomial draws}
\usage{
draw_multinom(bins, bin_sampled, size_sampled, n_draw, seedn = NULL)
}
\arguments{
\item{bins}{Matrix. Contains probabilities for multinomial draws}

\item{bin_sampled}{Vector. Column indices of bins to use for each draw}

\item{size_sampled}{Vector. Sizes to use for each draw}

\item{n_draw}{Integer. Number of random samples to take}

\item{seedn}{Boolean. Seed number}
}
\value{
Matrix with each column containing a random sample from the multinomial.
}
\description{
Draws random samples from a multinomial distribution. Takes \code{n_draw} draws 
from a multinomial with parameter p (probability) and k (size). The parameter p 
is stored in the columns of matrix \code{bins}. The column to select for each 
draw is given in the vector \code{bin_sampled}, which should be the same size as 
\code{n_draw}. The sizes of the draws k are given in the vector \code{size_sampled}, 
which also needs to be of length \code{n_draws}.
}
