% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{e_step}
\alias{e_step}
\title{E Step}
\usage{
e_step(x, k, mu, sgma, tau, semisup = FALSE, labels = NULL)
}
\arguments{
\item{x}{Matrix or data frame. Data observations with n rows and p columns}

\item{k}{Integer. Number of clusters}

\item{mu}{List of numeric vectors. Contains averages}

\item{sgma}{List. List of numeric vectors containing diagonal elements of sigma}

\item{tau}{Numeric. A vector of length \code{2}; the mixing parameter}

\item{semisup}{Boolean. Semi-supervised EM}

\item{labels}{Integer. If \code{semisup}, then give the labels}
}
\value{
Z numeric matrix with membership probabilities
}
\description{
E Step
}
