% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{est_dir}
\alias{est_dir}
\title{Estimate parameters of Dirichlet}
\usage{
est_dir(x, a, N = 1, rho = 0.1, eps = 1e-10, max_iter = 1e+05)
}
\description{
x are the observed data values of dirichlet
a are the alpha parameters of dirichlet
x and a must be same length
}
