% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{estim_prop_nnls}
\alias{estim_prop_nnls}
\title{Estimate percent of cell types in droplets}
\usage{
estim_prop_nnls(x, show_progress = TRUE, k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{show_progress}{Boolean indicating whether to show a progress
bar.}

\item{k_init}{Run EM on the \code{k_init} initialization(s). 
If NULL (default), run on all \code{k_init} initializations.}
}
\value{
An SCE object, with the prop slot containing a
 cluster by droplet matrix of cell type proportions
}
\description{
Use results from DIEM to estimate the 
cell type proportions within each droplet using nnls
}
