% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{estimate_db}
\alias{estimate_db}
\title{Estimate percent of debris in droplets}
\usage{
estimate_db(x, deb_clust = c(1), name = "pct.debris")
}
\arguments{
\item{x}{An SCE object.}

\item{deb_clust}{A numeric vector of debris clusters. Contains 
only cluster 1 by default, and should always contain cluster 1.}

\item{name}{Column name in test_data to store debris percent in.}
}
\value{
An SCE object
}
\description{
Given the debris clusters in deb_clust, estimate the percent 
of debris within each droplet. Store the percent in the 
test_data slot.
}
