% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{estimate_db_score}
\alias{estimate_db_score}
\title{Estimate debris score per droplet}
\usage{
estimate_db_score(x, thresh_mean = 0.5, thresh_p = 0.05,
  dbr_clust = NULL, verbose = TRUE, k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{thresh_mean}{Threshold for cluster mean debris score. Clusters 
with a mean above this value are considered debris clusters}

\item{thresh_p}{P-value threshold for calculating differential 
expression between droplets in the debris and non-debris clusters.}

\item{dbr_clust}{Specifies additional clusters to set as debris. Cluster 
1 is always included. Use this only if you know what you are doing.}

\item{verbose}{Verbosity, indicating whether to show a progress
bar.}

\item{k_init}{Run EM on the \code{k_init} initialization(s). 
If NULL (default), run on all \code{k_init} initializations.}
}
\value{
An SCE object
}
\description{
Estimate debris score per droplet
}
