% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{eval_iter}
\alias{eval_iter}
\title{Evaluate results from iteration}
\usage{
eval_iter(x, thresh = 0.05, verbose = FALSE)
}
\arguments{
\item{x}{SCE object.}

\item{thresh}{Numeric. The percent change (a numeric from 0 to 1).}

\item{verbose}{Boolean. Verbosity.}
}
\value{
A boolean indicating whether the change in nuclei droplets is under the threshold.
}
\description{
Evaluate results from iteration
}
