% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{fill_counts}
\alias{fill_counts}
\title{Fill information from raw counts}
\usage{
fill_counts(x)
}
\arguments{
\item{x}{An SCE object with x@counts data present}
}
\description{
Fill information from raw counts
}
