% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{fill_dropl_info}
\alias{fill_dropl_info}
\title{Insert pi and calls to dropl_info}
\usage{
fill_dropl_info(x)
}
\arguments{
\item{x}{SCE. SCE object}
}
\value{
Character vector with droplet IDs of called targets
}
\description{
Insert pi and calls to dropl_info
}
