% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter out lowly expressed genes}
\usage{
filter_genes(x, cpm_thresh = 0, verbose = FALSE)
}
\arguments{
\item{x}{An SCE object.}

\item{cpm_thresh}{CPMs must be greater than this value. This is 0 by 
default to include all expressed genes.}

\item{verbose}{verbosity}
}
\value{
An SCE object
}
\description{
This function removes genes that are lowly expressed. 
The droplets are split into the test set and debris set, and the Counts 
Per Million mapped reads (CPM) are calculated for each. Then, only genes 
that have a CPM greater than \code{cpm_thresh} in across the entire 
data set are kept.
}
