% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clust.R
\name{find_communities}
\alias{find_communities}
\title{Get louvain clusters}
\usage{
find_communities(knn_df, nrun = 1, method = "louvain")
}
\value{
Numeric vector of memberships
}
\description{
Get louvain clusters
}
