% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fix_tails}
\alias{fix_tails}
\title{Fix labels of lower and upper count droplets}
\usage{
fix_tails(x, pct = 0.05)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{log_base}{Numeric. The base of log to take}
}
\value{
An SCE object
}
\description{
Fix labels of the upper and lower tails droplets to nuclei and background, respectively.
The top and bottom \code{pct} are then fixed during EM.
}
