% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{fix_z}
\alias{fix_z}
\title{Fix posterior probabilities of labeled to 1}
\usage{
fix_z(Z, labs = NULL)
}
\arguments{
\item{Z}{A droplet by cluster matrix of the posterior probability a 
droplet belongs to cluster k. Rows must sum to 1 and entries must 
be between 0 and 1.}

\item{labs}{A droplet-length vector containing labels of the droplets. 
Non-zero values indicate unlabeled droplets.}
}
\value{
The droplet by cluster matrix with labeled droplets fixed 
 to ther corresponding cluster.
}
\description{
Fix posterior probabilities of labeled to 1
}
