% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{fraction_log}
\alias{fraction_log}
\title{Fraction of logs}
\usage{
fraction_log(x)
}
\arguments{
\item{x}{numeric vector. Values in log space}
}
\value{
numeric vector of relative fractions of logs
}
\description{
Given log values x1, .., xn, calculate for each value, the relative fraction
e.g. x1 / (x1 + ... + xn)
}
