% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{gene_data}
\alias{gene_data}
\title{Return the gene data from an SCE object}
\usage{
gene_data(x, min_droplets = 1)
}
\arguments{
\item{x}{An SCE object.}

\item{min_droplets}{Minimum number of droplets detected per gene}
}
\value{
A data frame
}
\description{
Return the data frame stored in the slot \code{gene_data}. This 
contains data such as the number of droplets the gene is detected in. 
The parameter \code{min_droplets} removes genes (rows) taht are 
detected in less than this number of droplets.
}
