% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_alpha}
\alias{get_alpha}
\title{LOO optimization}
\usage{
get_alpha(counts, Z, eps = 1e-04, max_loo = 500, psc = 1e-10,
  debug = FALSE)
}
\description{
counts is a droplet by gene matrix
sizes is total droplet size
a is current value of alpha parameter, gene by k matrix
z is current membership
}
