% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_alpha_mult}
\alias{get_alpha_mult}
\title{Get MLE of multinomial alphas}
\usage{
get_alpha_mult(counts, Z, alpha_prior = 0, add = 0, add_to = 1,
  psc = 1e-10, verbose = FALSE)
}
\arguments{
\item{counts}{A gene by droplet matrix}

\item{Z}{A droplet by cluster matrix of the posterior probability a 
droplet belongs to cluster k. Rows must sum to 1 and entries must 
be between 0 and 1.}

\item{alpha_prior}{Add a non-informative prior by adding a count of
\code{alpha_prior} to all genes in all clusters.}

\item{add}{A number, or weight, to add to the column sums of the 
posterior probabilities. This number is added to the \code{add_to} index.}

\item{add_to}{The index to add the number \code{add} to.}

\item{psc}{The pseudocount to add to the final MLE estimate. This avoids 
inappropriate alpha values of 0.}

\item{verbose}{Verbosity.}
}
\description{
Given the count data and posterior probabilities (membership), 
update alpha by finding the MLE. Weighted row sum
}
