% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_bg_cand}
\alias{get_bg_cand}
\title{Specify background droplets and candidate droplets}
\usage{
get_bg_cand(x, bg_min, bg_max, top_n_cand)
}
\arguments{
\item{x}{SCE. An SCE object}

\item{bg_min}{Numeric. Minimum number of counts for a droplet to be considered background}

\item{bg_max}{Numeric. Maximum number of counts for a droplet to be considered background}

\item{top_n_cand}{Numeric. Specifies the top number of droplets by expression to considered candidates}
}
\value{
SCE object with bg_info filled, a list with components
\item{Background}{Character vector of column names of x that are considered background droplets}
\item{Candidate}{Character vector of column names of x that are considered candidate droplets}
\item{Labels}{Numeric vector for droplets specifying 0 for candidate and 1 for background}
}
\description{
This function returns the barcodes that are considered originating from background, and those that are
considered candidate targets (cells/nuclei). Barcodes are considered background if their 
total read/UMI count is greater than or equal to \code{bg_min} and less than or equal to \code{bg_max}. 
Candidate cells/nuclei are the top \code{top_n_cand} droplets ranked by read counts. If the 
top \code{top_n_cand} droplets includes those with read/UMI counts < \code{bg_max}, the \code{bg_max} 
parameter is lowered so that \code{top_n_cand} candidates are always output.
}
