% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_bgscore}
\alias{get_bgscore}
\title{Get background score for each cell}
\usage{
get_bgscore(x)
}
\arguments{
\item{x}{SCE. SCE object.}
}
\value{
SCE object.
}
\description{
Background score is calculated as the percentage of reads mapping to a 
DE gene enriched in the background droplets. Differentially expressed genes 
must be calculated beforehand.
}
