% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_candd_counts}
\alias{get_candd_counts}
\title{Get expression matrix of candidate counts}
\usage{
get_candd_counts(x)
}
\arguments{
\item{x}{SCE.}
}
\value{
\code{\link[Matrix]{Matrix}} object
}
\description{
Get expression matrix of candidate counts
}
