% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{get_clust_dat}
\alias{get_clust_dat}
\title{Estimate proportions of each cluster given counts}
\usage{
get_clust_dat(counts, means, w_init, lrate = 0.1, eps = 0.001,
  threads = 1, verbose = TRUE)
}
\arguments{
\item{means}{Means for each cluster under the multinomial model.}

\item{w_init}{Initializing values for the proportions.}

\item{lrate}{Learning rate.}

\item{eps}{Threshold for convergence.}

\item{threads}{Number of threads.}

\item{verbose}{Verbosity.}

\item{x}{A sparse count matrix.}
}
\value{
A matrix containing fractions of each cluster.
}
\description{
Estimate the proportion of each cell type within a droplet using 
the multinomial parameters of the fitted model.
}
