% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{get_clust_f}
\alias{get_clust_f}
\title{Estimate proportions of each cluster for each droplet}
\usage{
get_clust_f(x, droplets = NULL, lrate = 0.1, eps = 0.001,
  threads = 1, verbose = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{droplets}{Specify droplets to calculate proportions for. If
not specified, calculates for all test droplets, which can be slow.}

\item{lrate}{Learning rate.}

\item{eps}{Threshold for convergence.}

\item{threads}{Number of threads}

\item{verbose}{Verbosity.}
}
\value{
A matrix containing fractions of each cluster.
}
\description{
Estimate the proportion of each cell type within a droplet using 
the multinomial parameters of the fitted model.
}
