% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_col_props}
\alias{get_col_props}
\title{Get proportion of counts in \code{rows}in x.}
\usage{
get_col_props(x, rows)
}
\arguments{
\item{x}{Matrix. Matrix object.}

\item{rows}{Character. rownames in x to calculate proportions of.}
}
\value{
Numeric vector.
}
\description{
Get proportion of counts in \code{rows}in x.
}
