% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_de_genes_t}
\alias{get_de_genes_t}
\title{Get DE genes between high and low read count droplets using t-test}
\usage{
get_de_genes_t(x, max_drop = 10000, fc_thresh = 2, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{max_drop}{Numeric. Maximum number of droplets to use for t-test}

\item{fc_thresh}{Numeric. Only genes that have this fold-change threshold between low and high
droplets will be considered as differentially expressed. Set to 2 by default.}

\item{verbose}{Boolean.}
}
\value{
SCE object with \code{gene_info[,"p"]}, \code{gene_info[,"fc"]}, and \code{deg} slots filled. 
The \code{p} column contains the background - candidate difference in proportions.
The \code{fc} column contains the fold change: background over high read count
The \code{deg} slot contains the names of the genes output as DE.
}
\description{
Find differentially expressed genes between low and high read count droplets. Sums
gene expression across all cells within the limits specified in the SCE object, 
divides by the total number of reads in each, and calculates the difference between 
the 2. The top \code{n_genes} genes sorted by absolute difference in proportion are 
designated as DE.
}
