% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{get_diff_thresh}
\alias{get_diff_thresh}
\title{Set count cutpoint for DE testing between low and high count droplets}
\usage{
get_diff_thresh(x, top_n = 10000, diff_thresh = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{diff_thresh}{Numeric. Manually initialize the DE cutpoint.}

\item{verbose}{Boolean. Verbosity.}

\item{log_base}{Numeric. Base of log to take.}
}
\value{
SCE object.
}
\description{
In order to find genes enriched in the background or the signal, we need to 
determine how to group droplets for DE. This function initializes a cut point 
to decide which droplets are assigned to either the high or low group. This number 
is determined by dividing the maximum droplet count by \code{log_base}. 
Alternatively, the you can set the cutpoint threshold manually with \code{de_cutpoint}.
}
