% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_exp_llk}
\alias{get_exp_llk}
\title{Expected log likelihood of multivariate normal under 1 or more latent clusters.}
\usage{
get_exp_llk(x, mu, sgma, tau, Z)
}
\arguments{
\item{x}{matrix or data frame. Observations with which to calculate likelihood under parameters.}

\item{mu}{list of numeric vectors. Averages for each of the k clusters.}

\item{sgma}{list of numeric vectors. Covariance for each of the k clusters.}

\item{tau}{numeric vector. Mixing parameters for each of the k clusters.}

\item{Z}{matrix or data frame. Latent membership probabilities. n x k matrix.}
}
\value{
numeric value of the expected log likelihood under k clusters.
}
\description{
Calculates expected log likelihood of multivariate distribution given data and parameters. 
For each of k clusters, calculate log likelihood, multiply by its mixing parameter tau, and 
sum across clusters.
}
