% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_gene_pct}
\alias{get_gene_pct}
\title{Get percent of reads aligning to given genes}
\usage{
get_gene_pct(x, genes, name)
}
\arguments{
\item{x}{SCE.}

\item{genes}{Character. Genes to calculate percentage of in counts}

\item{name}{Character. Column name to place in dropl_info}
}
\value{
SCE object
}
\description{
Get percent of reads aligning to given genes
}
