% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_llk_ss}
\alias{get_llk_ss}
\title{Semisupervised log likelihood of multivariate normal with labeled and unlabeled data}
\usage{
get_llk_ss(x, mu, sgma, tau, labels)
}
\arguments{
\item{x}{matrix or data frame. Observations with which to calculate likelihood under parameters.}

\item{mu}{list of numeric vectors. Averages for each of the k clusters.}

\item{sgma}{list of numeric vectors. Covariance for each of the k clusters.}

\item{tau}{numeric vector. Mixing parameters for each of the k clusters.}

\item{labels}{numeric vector. Labels of observations. 0 for unknown.}
}
\value{
numeric value of the observed log likelihood under k clusters.
}
\description{
Calculates log likelihood of multivariate distribution given data and parameters. 
For each of k clusters, calculate log likelihood, multiply by its mixing parameter tau, and 
sum across clusters.
}
