% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{get_log2fc}
\alias{get_log2fc}
\title{Run differential expression between high and low read count droplets}
\usage{
get_log2fc(x, cpm_thresh = 3, log2fc_thresh = 0, iteration = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{cpm_thresh}{Numeric. CPM treshold for expressed genes in both low and high groups.}

\item{log2fc_thresh}{Numeric. Call DE genes with log2 fold change above or below this threshold.}

\item{verbose}{Boolean.}
}
\value{
DE object.
}
\description{
Find differentially expressed genes between low and high read count droplets. Groups droplets into
low or high count groups. The count threshold for separating low and high is determined by taking the 
log10 of the maximum droplet count, subtracting 1, and calculating 10 to the power of this number. In other 
words, the thresold is found by dividing the max count by 10.
}
