% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_mt_malat1}
\alias{get_mt_malat1}
\title{Get percent of reads aligning to MT genome and MALAT1 gene}
\usage{
get_mt_malat1(x)
}
\arguments{
\item{x}{SCE.}
}
\value{
SCE object
}
\description{
Places percent of reads aligning to MT genome and MALAT1 gene in 
\code{x@dropl_info}, under columns MALAT1 and MT_PCT. The gene 
names in the rows of \code{x@counts} should be MALAT1 and only mitochondrial 
genes should start with MT-.
}
