% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_pi}
\alias{get_pi}
\title{Get pi}
\usage{
get_pi(x)
}
\arguments{
\item{x}{SCE. SCE object.}
}
\value{
An SCE object.
}
\description{
This function calculates pi_low and pi_high, defined as the inner products of 
normalized gene expression and log fold changes for DE genes enriched in the low (background-enriched) 
and high (target-enriched) count droplets, respectively.
}
