% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_removed_ids}
\alias{get_removed_ids}
\title{Return IDs of removed droplets}
\usage{
get_removed_ids(x, min_genes = 200)
}
\arguments{
\item{x}{An SCE object.}

\item{min_genes}{The minimum number of genes detected for a 
droplet to be output. Useful if a threshold was used to call 
targets, as droplets below would never be considered. By 
default is set at 200.}
}
\value{
A character vector with the called droplet IDs.
}
\description{
Return the droplet IDs in the test set that have been removed 
by DIEM. 
The function \code{\link{call_targets}} must be run beforehand.
}
