% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{get_snn}
\alias{get_snn}
\title{Get shared nearest neighbors}
\usage{
get_snn(x, nn = 30, kt = 3, min_counts = 10, droplets.use = NULL,
  weighted = TRUE, algorithm = "kd_tree", verbose = FALSE)
}
\description{
Get shared nearest neighbors
}
