% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{get_var_genes}
\alias{get_var_genes}
\title{Get variable genes}
\usage{
get_var_genes(x, cpm_thresh = 10, ng = 2000, lss = 0.3,
  verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{lss}{Numeric. The span paramater for loess function.}

\item{verbose}{Boolean.}

\item{min_mean}{Numeric. Minimum mean of counts to consider.}

\item{nf}{Numeric. Output top nf variable features.}
}
\value{
SCE object with \code{vg} and \code{vg_info} slots filled for variable genes
and variable gene info. \code{vg_info} is a data frame with log mean, log var, 
fitted log var values, and residuals. \code{vg} is a vector of genes
}
\description{
Finds top \code{nf} variable genes by running a loess regression 
on log(variance) and log(mean). Residuals are used to rank 
features, and the top \code{nf} are output.
}
