% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmap_pi_genes}
\alias{heatmap_pi_genes}
\title{Heatmap of pi-gene expression correlations}
\usage{
heatmap_pi_genes(x, top_n = 15, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless ret=TRUE then a ggplot
}
\description{
Heatmap of pi-gene expression correlations
}
