% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{init_best}
\alias{init_best}
\title{Initialize best from random starts}
\usage{
init_best(X, K, hyperp, n_start = 100, n_iter = 3, labs = NULL,
  verbose = TRUE)
}
\description{
Initialize best from random starts
}
