% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{init_de_cutpoint}
\alias{init_de_cutpoint}
\title{Set count cutpoint for DE testing between low and high count droplets}
\usage{
init_de_cutpoint(x, log_base = 100, de_cutpoint = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{log_base}{Numeric. Base of log to take.}

\item{de_cutpoint}{Numeric. Manually set the DE cutpoint.}

\item{verbose}{Boolean. Verbosity.}
}
\value{
SCE object.
}
\description{
In order to find genes enriched in the background or the nucleus, we need to 
determine how to group droplets for DE. This function finds a proper cut point 
to decide which droplets are assigned to either the high or low group. This number 
is determined by dividing the maximum droplet count by \code{log_base}. 
Alternatively, the you can set the cutpoint threshold manually with \code{de_cutpoint}.
}
