% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{init_param}
\alias{init_param}
\title{Initialize parameters for EM}
\usage{
init_param(x, k, labels = NULL)
}
\arguments{
\item{x}{Matrix or data frame. Data observations with n rows and p columns}

\item{k}{Integer. Number of clusters}
}
\value{
Parameters in a list with components
\item{mu}{List of length k, with each element a numeric vector of mus of length p}
\item{sgma}{List of length k, with each element a numeric vector of covariance}
\item{tau}{Numeric vector of length k}
}
\description{
Initialize parameters for EM
}
