% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{init_tails}
\alias{init_tails}
\title{Initialize EM parameters for k=2 low and high groups}
\usage{
init_tails(counts, labels)
}
\value{
List with mu0 a p x 2 matrix and mc0 a length 2 numeric
}
\description{
Initialize the means of the two multinomial groups. Labels=2 are debris
}
