% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{init_test_set}
\alias{init_test_set}
\title{Subset droplets for EM testing}
\usage{
init_test_set(x, log_base = 10, top_n = NULL, min_counts = NULL)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{log_base}{Numeric. Base of log to take.}

\item{top_n}{Numeric. Only the top \code{top_n} droplets ranked by total counts are classified.}

\item{min_counts}{Integer. Only droplets with at least this number of total counts are classified.}
}
\value{
An SCE object.
}
\description{
Sets the threshold for testing droplets and subsets counts matrix. Only droplets with a total number 
of counts of at least \code{min_counts} are included in the semi-supervised EM and assigned as either 
background or target. By default, this number is calculated by taking 2 orders of magnitude lower than
the maximum droplet count in the data. The order of magnitude is set by \code{log_base}. In other words, 
this divides the max droplet count by the square of \code{log_base} and sets this as the minimum count. 
The minimum number of counts can be set using the \code{top_n} parameter instead, so that only the 
\code{top_n} droplets ranked by total counts are used in testing. Another way is to directly set the 
minimum counts through \code{min_counts}. The count matrix is subset to droplets having total counts greater 
than this threshold.
}
