% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clust.R
\name{initialize_clusters}
\alias{initialize_clusters}
\title{Initialize clusters for unlabeled data}
\usage{
initialize_clusters(x, method = "louvain", cor_thresh = 0.95)
}
\value{
Numeric vector of memberships
}
\description{
Initialize clusters for unlabeled data
}
