% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{low_high_cor_plot}
\alias{low_high_cor_plot}
\title{Correlation plot of low vs high expression}
\usage{
low_high_cor_plot(x, scale_factor = 1, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Correlation plot of low vs high expression
}
