% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{m_step}
\alias{m_step}
\title{M Step}
\usage{
m_step(x, Z)
}
\arguments{
\item{x}{Matrix or data frame. Data observations with n rows and p columns}

\item{Z}{}
}
\value{
list with updated mu, sigma, and tau parameters
}
\description{
M Step
}
