% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb_small}
\alias{mb_small}
\title{Single-nucleus RNA-seq of mouse brain}
\format{An SCE object with 1,000 rows (genes) and 1,386 droplets (columns)}
\source{
\url{https://www.10xgenomics.com}
}
\usage{
mb_small
}
\description{
A dataset containing a subset of the publicly available 
10X single-nucleus RNA-seq data set of freshly dissected 
mouse brain tissue.
}
\keyword{datasets}
