% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{merge_diem}
\alias{merge_diem}
\title{Merge and re-cluster multiple DIEM objects}
\usage{
merge_diem(x, use_var_genes = FALSE, eps = 1e-08, max_iter = 1000,
  hcut = 0.05, sep = "_", name = "Merged", verbose = TRUE)
}
\arguments{
\item{x}{A list of SCE objects.}
}
\value{
An SCE object.
}
\description{
Merge and re-cluster multiple DIEM objects
}
