% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{mvn_logllk_diag}
\alias{mvn_logllk_diag}
\title{Log-likelihod of multivariate normal with diagonal covariance}
\usage{
mvn_logllk_diag(x, mu, sgma)
}
\arguments{
\item{x}{numeric vector or matrix. The data observations}

\item{mu}{numeric vector. Mu average of each variable in the multivariate normal}

\item{sgma}{numeric vector or diagonal matrix. If matrix, extracts diagonal only.}
}
\value{
numeric value of the log likelihood of the given observation x under
 parameters mu and sgma
}
\description{
Calculate log-likelihood of sample x given mu and diagonal sigma.
}
