% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{norm_counts}
\alias{norm_counts}
\title{Normalize counts of a sparse matrix}
\usage{
norm_counts(counts, scale_factor = 10000, logt = TRUE)
}
\arguments{
\item{counts}{Sparse Matrix}

\item{scale_factor}{Numeric. A scaling factor to multiply values after division by total count.
Default is 1.}

\item{logt}{Boolean. Log transform after normalizing for columns to sum to 1 and multiplying by scaling factor.
Default is TRUE}
}
\value{
Sparse Matrix
}
\description{
Normalize counts of a sparse matrix
}
