% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{normalize}
\alias{normalize}
\title{Normalize (total count and log) raw count data}
\usage{
normalize(x, sf = 10000)
}
\arguments{
\item{x}{SCE. SCE}

\item{sf}{Numeric. A scaling factor to multiply values after division by total count}
}
\value{
A matrix with normalized expression data
}
\description{
Raw expression counts are divided by total counts per droplet, 
multiplied by scaling factor, and log1p transformed. Same 
normalization style as Seurat. This normalizes the counts in 
\code{raw} slot and places them in the @norm slot.
}
