% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{params_frm_clust}
\alias{params_frm_clust}
\title{Get params from cluster assignments}
\usage{
params_frm_clust(x, clusts, genes.use, model = "mltn", alpha_prior = 0,
  pi_prior = 0, psc = 1e-10, threads = 1)
}
\arguments{
\item{x}{An SCE object.}

\item{clusts}{Integer vector of cluster assignments. Names designate 
the droplets}

\item{genes.use}{Character vector of gene names}
}
\description{
Get params from cluster assignments
}
