% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pc_bgscore_plot}
\alias{pc_bgscore_plot}
\title{Background score-PC plot}
\usage{
pc_bgscore_plot(x, title = "", pc = 1, return = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{title}{Character}

\item{return}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatter plot of background scores vs PCs.
}
