% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{pi_init_tails}
\alias{pi_init_tails}
\title{Initialize EM parameters for k=2 low and high groups}
\usage{
pi_init_tails(x, pct = 0.05)
}
\arguments{
\item{x}{SCE. SCE object with pi calculated}

\item{pct}{Numeric. Percent of droplets at each tail to fix. A number from 0 to 1.}
}
\value{
A list with 3 components
\itemize{
\item{"mu0"}{"A list of size 2, with each element being a size 2 vector containing the means of pi_l and pi_h."}
\item{"sgma0"}{"A list of size 2, with each element being a size 2 vector containing the variances of pi_l and pi_h."}
\item{"tau0"}{"A numeric of size 2, containing the tau mixing coefficients."}
}
The first and second element of the lists of mu0 and sgma0 are for the high and low groups respectively
}
\description{
Initialize the means and variances of the 2 Gaussians, as well as tau, 
the mixing parameter.
}
