% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clust}
\alias{plot_clust}
\title{Scatterplot of cluster means for indicated variables}
\usage{
plot_clust(sce, feat_x = "score.debris", feat_y = "n_genes",
  log_x = TRUE, log_y = FALSE, ret = FALSE)
}
\arguments{
\item{sce}{An SCE object.}

\item{feat_x}{Feature to plot on the x-axis.}

\item{feat_y}{Feature to plot on the y-axis.}

\item{log_x}{Boolean indicating whether to log transform the x-axis.}

\item{log_y}{Boolean indicating whether to log transform the y-axis.}

\item{ret}{A logical specifying whether to return the ggplot object 
or just print it out.}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Features should be a column name from droplet_data
}
