% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_data}
\alias{plot_data}
\title{Scatterplot of features from meta data or gene expression}
\usage{
plot_data(sce, feat_x = "total_counts", feat_y = "n_genes",
  log_x = FALSE, log_y = FALSE, color = NULL, alpha = 0.1,
  data_type = "test", ret = FALSE)
}
\arguments{
\item{sce}{An SCE object.}

\item{feat_x}{Variable of x axis to plot.}

\item{feat_y}{Varibale of y axis to plot.}

\item{log_x}{Boolean indicating whether to log transform the y-axis.}

\item{log_y}{Boolean indicating whether to log transform the y-axis.}

\item{color}{Color points by values in this column of test_data, such 
as Call.}

\item{alpha}{A numeric value controlling the transparency of the points. 
From 0 (transparent) to 1 (no transparency).}

\item{data_type}{Either test or all, specifying the droplets to plot}

\item{ret}{A logical specifying whether to return the ggplot object 
or just print it out.}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Plot variables from meta data or from gene expression. \code{feat_x} 
will be plotted on the x-axis, while \code{feat_y} will be plotted on 
on the y-axis. These variables can come from test_data or the 
normalized gene expression values of a features.
}
