% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dbr_score}
\alias{plot_dbr_score}
\title{Scatterplot of debris score against feature}
\usage{
plot_dbr_score(sce, feature = "n_genes", log_y = TRUE, color = NULL,
  alpha = 0.1, ret = FALSE)
}
\arguments{
\item{sce}{An SCE object.}

\item{feature}{Plot debris score against this feature. This can 
be a column name in the test_data slot or a gene. Default is 
number of genes detected}

\item{log_y}{Boolean indicating whether to log transform the y-axis.}

\item{color}{Color points by values in this column of test_data, such 
as Call.}

\item{alpha}{A numeric value controlling the transparency of the points. 
From 0 (transparent) to 1 (no transparency).}

\item{ret}{A logical specifying whether to return the ggplot object 
or just print it out.}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Scatterplot of debris score against feature
}
