% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deltas}
\alias{plot_deltas}
\title{Scatterplot of delta for each iteration}
\usage{
plot_deltas(sce, eps = 1e-04, ret = FALSE)
}
\arguments{
\item{sce}{An SCE object.}

\item{eps}{Threshold for delta, plotted as dashed red line}

\item{ret}{A logical specifying whether to return the ggplot object 
or just print it out.}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Features should be a column name from droplet_data
}
