% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_iter}
\alias{plot_iter}
\title{Plot through iterations, return a list of ggplot2 objects}
\usage{
plot_iter(x, FUN, ret = FALSE, ...)
}
\arguments{
\item{x}{SCE. SCE object}

\item{FUN}{character string. Name of function for plotting. Can be one of 
of the following:
\itemize{
   \item de_cor_plot
   \item llk_fraction_plot
   \item plot_pi
   \item plot_mvn_contour
}}

\item{ret}{Logical. If TRUE, return the list of ggplots, else, print them}
}
\value{
Nothing
}
\description{
Plot through iterations, return a list of ggplot2 objects
}
