% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_iter_gif}
\alias{plot_iter_gif}
\title{Plot through iterations in a gif}
\usage{
plot_iter_gif(x, FUN, prefix = "iter", ...)
}
\arguments{
\item{x}{SCE. SCE object}

\item{FUN}{character string. Name of function for plotting. See \code{plot_iter} for more details.}

\item{prefix}{character string. Prefix of GIF output file name.}
}
\value{
Nothing
}
\description{
Plot through iterations in a gif
}
