% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_kld}
\alias{plot_kld}
\title{Barplot of KL-divergence between low-count and 1. debris and 2. signal means}
\usage{
plot_kld(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Barplot of KL-divergence between low-count and 1. debris and 2. signal means
}
