% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mvn_contour}
\alias{plot_mvn_contour}
\title{Contour plot of MVN densities for the 2 groups}
\usage{
plot_mvn_contour(x, length.out = 100, iteration = NULL, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Contour plot of MVN densities for the 2 groups
}
