% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi}
\alias{plot_pi}
\title{Scatterplot of pi}
\usage{
plot_pi(x, color = NULL, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{return}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatterplot of pi
}
