% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_contour}
\alias{plot_pi_contour}
\title{Contour plot of pi values}
\usage{
plot_pi_contour(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Contour plot of pi values
}
