% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_iter_gif}
\alias{plot_pi_iter_gif}
\title{Plot of pi values through iterations in a gif}
\usage{
plot_pi_iter_gif(x, prefix = "pi_iter")
}
\arguments{
\item{x}{SCE. SCE object}
}
\value{
Nothing
}
\description{
Plot of pi values through iterations in a gif
}
