% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_marginal}
\alias{plot_pi_marginal}
\title{Scatterplot of pi with density plots on the margin}
\usage{
plot_pi_marginal(x, alpha = 0.05, pdfname = NULL, w = 7, h = 7)
}
\arguments{
\item{x}{SCE. SCE object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatterplot of pi with density plots on the margin
}
