% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_tails}
\alias{plot_pi_tails}
\title{Scatterplot of pi colored by tails}
\usage{
plot_pi_tails(x, pct = 0.05, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatterplot of pi colored by tails
}
