% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_top}
\alias{plot_pi_top}
\title{Scatterplot of pi coloring top expressed droplets}
\usage{
plot_pi_top(x, top_n = 1000, color = NULL, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{return}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatterplot of pi coloring top expressed droplets
}
