% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umi_gene}
\alias{plot_umi_gene}
\title{Scatterplot of genes vs. UMI counts, colored by posterior probability}
\usage{
plot_umi_gene(x, color = "CleanProb",
  color_name = "Probability\\nClean", palette = "PuBuGn",
  alpha = 0.1, ret = FALSE)
}
\arguments{
\item{x}{An SCE object.}

\item{color}{Column name of droplet_data to color points by.}

\item{color_name}{Title of the color legend.}

\item{palette}{Name of the palette to use from RColorBrewer;
see \code{\link[ggplot2]{scale_color_distiller}}.}

\item{alpha}{A numeric value controlling the transparency of the points. 
From 0 (transparent) to 1 (no transparency).}

\item{ret}{A logical specifying whether to return the ggplot object 
or just print it out.}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Scatterplot of genes vs. UMI counts, colored by posterior probability
}
