% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umi_gene_pp}
\alias{plot_umi_gene_pp}
\title{Scatterplot of genes vs. UMI counts, colored by posterior probability}
\usage{
plot_umi_gene_pp(x, alpha = 0.1, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{alpha}{SCE. Transparency of points. 0 (transparent) to 1 (no transparency)}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Scatterplot of genes vs. UMI counts, colored by posterior probability
}
