% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pp_plot}
\alias{pp_plot}
\title{Plot posterior probability against pi_low - pi_high}
\usage{
pp_plot(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Plot posterior probability against pi_low - pi_high
}
