% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_10x}
\alias{read_10x}
\title{Read 10X output}
\usage{
read_10x(path)
}
\arguments{
\item{path}{Character. File path prefix to 10X output.}
}
\value{
expr Sparse matrix. Expression counts from 10X cell ranger
 with cells in the columns and genes in the rows.
}
\description{
Read output from 10X into a sparse matrix. Given a 10X output path, 
containing the output files from 10X CellRanger. These should be the 
raw output files \code{barcodes.tsv} \code{genes.tsv} \code{matrix.mtx} 
if using CellRanger v2, or \code{barcodes.tsv.gz} \code{features.tsv.gz} 
\code{matrix.mtx.gz} if using CellRanger v3.
}
\examples{
counts <- read_10x("mouse_nuclei_2k/raw_gene_bc_matrices/mm10/")

}
