% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{remove_debris}
\alias{remove_debris}
\title{Remove debris reads from droplets}
\usage{
remove_debris(x, verbose = FALSE)
}
\value{
An SCE object
}
\description{
This function takes the raw counts matrix and removes the counts that 
are estimated to originate from debris. For each droplet, the percent 
of debris and cell type are estimated by non-negative least squares. 
Then, a debris droplet is calculated with same total read count as 
the droplet and multiplied by the debris proportion. The non-integer 
numbers are rounded down. Finally, this debris vector is subtracted 
from the droplet counts, keeping reads counts to non-negative.
}
