% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{run_diem_10x_sim}
\alias{run_diem_10x_sim}
\title{Run diem pipeline for 10X samples}
\usage{
run_diem_10x_sim(path_10X, verbose = FALSE, ...)
}
\value{
List of SCE objects.
}
\description{
Wrapper to run diem pipeline on 10X samples, returning \code{SCE} objects.
}
