% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{run_em}
\alias{run_em}
\title{Run EM on pi features}
\usage{
run_em(x, pct_tails = 0.05, min_iter = 5, max_iter = 1000,
  eps = 1e-10, verbose = TRUE)
}
\arguments{
\item{x}{SCE. SCE object with PCs.
score and the absolute value of the R coefficient must be greater than this threshold.}

\item{min_iter}{Integer. Minimum number of iterations.}

\item{max_iter}{Integer. Maximum number of iterations.}

\item{eps}{Numeric. Epsilon of log-likilhood change to call convergence.}

\item{verbose}{Boolean. Verbosity.}
}
\value{
SCE object with EM output in the \code{emo} slot. See \code{\link{run_mv_em_diag}} for details
}
\description{
Run expectation maximization (EM) using diagonal covariance on pi matrix in an SCE object with k=2 groups. 
EM is run \code{n_runs} times with random initializations, and parameters with 
the best log likelihood are output. Within EM, iterate a minimum of \code{min_iter} times and 
a maximum of \code{max_iter} times. The \code{eps} value gives the threshold of percentage change in 
log likelihood until convergence is reached.
}
