% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SCE-class}
\alias{SCE-class}
\alias{SCE}
\title{SCE}
\description{
The SCE class
}
\details{
This class is used to store the expression data from a 
single-cell RNA-seq experiment. It is the class that is used by 
DIEM.

Single Cell Expression object
}
