% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_cluster_set}
\alias{set_cluster_set}
\title{Set droplets for cluster initialization}
\usage{
set_cluster_set(x, cluster_n = 1000, order_by = "gene",
  verbose = FALSE)
}
\arguments{
\item{x}{An SCE object.}

\item{top_n}{Numeric value specifying the  droplets below \code{top_n} (ranked by total counts) 
are fixed to debris. Set to NULL to ignore.}

\item{min_counts}{A numeric value that specifies that all droplets below this count threshold are 
always fixed to debris.}

\item{top_thresh}{A numeric value specifying any droplets above this count threshold are 
fixed to nuclei. No droplets are fixed to nuclei by default.}
}
\value{
An SCE object.
}
\description{
Set droplets for cluster initialization
}
