% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{set_de_cutpoint}
\alias{set_de_cutpoint}
\title{Set count cutpoint for DE testing between low and high count droplets}
\usage{
set_de_cutpoint(x, de_cutpoint = NULL, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{de_cutpoint}{Numeric. Manually set the DE cutpoint.}

\item{verbose}{Boolean. Verbosity.}
}
\value{
SCE object.
}
\description{
In order to find genes enriched in the background or the nucleus, we need to 
determine how to group droplets for DE. This function finds a proper cut point 
to decide which droplets are assigned to either the high or low group. This number 
is determined by taking the log10 of the maximum droplet count, subtracting 1, 
and calculating 10 to the power of this number (10^(log10(x)-1) where x is the max 
count. Alternatively, the you can set the cutpoint threshold manually.
}
