% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{set_de_groups}
\alias{set_de_groups}
\title{Set low and high droplet labels}
\usage{
set_de_groups(x, low = NULL, high = NULL)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{low}{Character. Droplet IDs to set as the low group.}

\item{high}{Character. Droplet IDs to set as the high group.}
}
\value{
SCE object.
}
\description{
Set the droplet IDs to be used as the low and high groups for 
DE testing. By default, the high group is set as the nuclei called 
by the previous iteration of DIEM, while the low group is set as 
everything else. If \code{low} and \code{high} are given, as in 
during initialization when nuclei haven't been called yet, the low 
and high groups are set to these for logFC calculation.
}
