% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{set_diff_genes}
\alias{set_diff_genes}
\title{Set differential genes between high and low read count droplets}
\usage{
set_diff_genes(x, n_genes = 100, cpm_thresh = 3, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{n_genes}{Numeric. Number of genes to set as differential.}

\item{cpm_thresh}{Numeric. CPM treshold for expressed genes in both low and high groups.}

\item{verbose}{Boolean.}
}
\value{
DE object.
}
\description{
Find differentially expressed genes between low and high read count droplets. Groups droplets into
low or high count groups. The count threshold for separating low and high is determined by taking the 
log10 of the maximum droplet count, subtracting 1, and calculating 10 to the power of this number. In other 
words, the thresold is found by dividing the max count by 10.
}
