% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{set_double}
\alias{set_double}
\title{Round x up to double precision value if R rounds x below the minimum number for double}
\usage{
set_double(x)
}
\description{
Round x up to double precision value if R rounds x below the minimum number for double
}
