% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_labels}
\alias{set_labels}
\title{Label upper count droplets}
\usage{
set_labels(x, log_base = 2)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{log_base}{Numeric. The base of log to take}
}
\value{
An SCE object
}
\description{
Fix labels of upper count droplets to target according to total counts in a droplet. The threshold 
is set by taking the log of the maximum droplet count, substracting by 1, and exponentiating by 
base of the log. In other words, this divides the max count by \code{log_base} and sets labels of 
droplets with counts greater than this to target.
}
