% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{set_limits}
\alias{set_limits}
\title{Set count and gene limits for background and target}
\usage{
set_limits(x, min_bg_count = 0, max_bg_count = 150,
  min_tg_count = 199, max_tg_count = Inf, min_tg_gene = 199,
  max_tg_gene = Inf, top_n_tg = NULL, top_n_bg = NULL)
}
\arguments{
\item{x}{SCE object}

\item{min_bg_count}{Numeric. Minimum number of counts for a droplet to be considered background}

\item{max_bg_count}{Numeric. Maximum number of counts for a droplet to be considered background}

\item{min_tg_count}{Numeric. Minimum number of counts for a droplet to be considered target}

\item{max_tg_count}{Numeric. Maximum number of counts for a droplet to be considered target}

\item{min_tg_gene}{Numeric. Minimum number of genes detected for a droplet to be considered target}

\item{max_tg_gene}{Numeric. Maximum number of genes detected for a droplet to be considered target}

\item{top_n_tg}{Numeric. Set the top n ranked by counts as the threshold for considering targets}

\item{top_n_bg}{Numeric. Set the top n ranked by counts as the threshold for considering background}
}
\value{
SCE object
}
\description{
Set the minimum and maximum number of counts for a droplet to be considered
either background or droplet. Also sets the limit of minimum or maximum number 
of detected genes that a droplet can be considered a target. On top of setting 
the number explicitly, the min_tg_count and max_bg_count can be set by ranking 
of droplet total counts. For example, you can set droplets above the 10,000th 
ranked droplet as candidate targets and the all droplets under the 20,000th 
ranked droplet as background. The parameters top_n_tg and top_n_bg override 
the numeric specifications of max_bg_count and min_tg_count.
}
