% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_min_counts}
\alias{set_min_counts}
\title{Set counts threshold for testing droplets}
\usage{
set_min_counts(x, top_n = NULL, min_counts = NULL)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{top_n}{Numeric. Only the top \code{top_n} droplets ranked by total counts are used for testing.}

\item{min_counts}{Integer. Only droplets with at least this number of total counts are used for testing.}
}
\value{
An SCE object with \code{min_counts} slot set
}
\description{
Sets the threshold for testing droplets. Only droplets with a total number of counts of at least
\code{min_counts} are included in the semi-supervised EM and assigned as either background or target. 
By default, this number is calculated by dividing the maximum droplet count by 100 in order to test 
the top 2 orders of magnitude in the data. 
The minimum number of counts can be set using the \code{top_n} parameter instead, so that only the 
\code{top_n} droplets ranked by total counts are used in testing. Another way is to directly set the 
minimum counts through \code{min_counts}.
}
