% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_bg}
\alias{sim_bg}
\title{Simulate background droplets with higher read counts}
\usage{
sim_bg(x, simf = 0.5, gma = 10, seedn = NULL, verbose = FALSE)
}
\arguments{
\item{x}{SCE.}

\item{simf}{Numeric. Generate number of candidates * \code{simf} random samples.}

\item{gma}{Numeric. Multiply p by this number to generate alpha for Dirichlet-multinomial random sampling.}

\item{seedn}{Numeric. Number to set seed.}

\item{verbose}{Boolean}
}
\value{
SCE object.
}
\description{
Given an SCE object with raw expression data containing droplets from background 
and targets (cells/nuclei), generate simulated background droplets with the same 
count distribution as candidate targets. Then replace the expression data with 
counts only from the simulated droplets and the candidate target droplets. Labels 
the simulated background droplets in the column "background" of \code{x@dropl_info}. 
Candidate targets are the droplets with total counts and genes detected as 
specified in the SCE object limits. The simulated 
background droplets are generated from a multinomial distribution, which is estimated 
from the background droplets. The background droplets are identified as those with
total reads counts in the range between \code{start_bin} and \code{end_bin}. By default, 
only 1 multinomial distribution is estimated. However, several distributions can 
be estimated from bins of \code{by_bin} width between \code{start_bin} and \code{end_bin}.
}
