% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{subset_DE}
\alias{subset_DE}
\title{Subset DE S4 object to genes in keep}
\usage{
subset_DE(x, keep)
}
\arguments{
\item{x}{SCE object.}

\item{keep}{Character vector. Names of genes to keep in DE object}
}
\value{
An SCE object
}
\description{
Subset DE S4 object to genes in keep
}
