% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{subset_dropls}
\alias{subset_dropls}
\title{Subset droplets in SCE object}
\usage{
subset_dropls(x, min_c = -Inf, max_c = Inf, slot_name = "raw")
}
\arguments{
\item{x}{SCE object. Must contain data in raw slot}

\item{min_c}{Numeric. Minimum row sum}

\item{max_c}{Numeric. Maximum row sum}

\item{slot_name}{Character. Slot of SCE subset, either raw or norm, or both}
}
\value{
x SCE object subsetted
}
\description{
Given thresholds min_c and max_c, subset droplets by read counts
Keep droplets where the count is greater than or 
equal to min_c and less than or equal to max_c.
}
