% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{subset_genes}
\alias{subset_genes}
\title{Subset genes in SCE object}
\usage{
subset_genes(x, min_c = -Inf, max_c = Inf)
}
\arguments{
\item{x}{SCE object. Must contain data in raw slot}

\item{min_c}{Numeric. Minimum gene counts}

\item{max_c}{Numeric. Maximum gene counts}
}
\value{
x SCE object subsetted
}
\description{
Given thresholds min_c and max_c, subset genes by mean counts.
Keep genes where the count is greater than or 
equal to min_m and less than or equal to max_m.
}
