% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{sum_log}
\alias{sum_log}
\title{Sum of logs}
\usage{
sum_log(x)
}
\arguments{
\item{x}{numeric vector. Values in log space}
}
\value{
numeric value of the sum of the logs
}
\description{
Given log values x1, .., xn, calculate the sum: 
log( sum( exp(x1) + ... + exp(xn) ) )
}
