% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{summary_results}
\alias{summary_results}
\title{Summarize results}
\usage{
summary_results(x)
}
\arguments{
\item{x}{SCE. SCE object with PCs and labels}
}
\value{
SCE object with summary stats in \code{other_info}
}
\description{
Create a data frame summarizing summary stats for target and background calls.
}
