% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{t_test_sparse}
\alias{t_test_sparse}
\title{Get variable genes}
\usage{
t_test_sparse(x_mean, x_var, x_n, y_mean, y_var, y_n)
}
\arguments{
\item{x_mean}{Numeric. Mean of group x.}

\item{x_var}{Numeric. Variance of group x.}

\item{x_n}{Numeric. Size of group x.}

\item{y_mean}{Numeric. Mean of group y.}

\item{y_var}{Numeric. Variance of group y.}

\item{y_n}{Numeric. Size of group y.}

\item{x}{SCE. SCE object.}

\item{min_mean}{Numeric. Minimum mean of counts to consider.}

\item{nf}{Numeric. Output top nf variable features.}

\item{lss}{Numeric. The span paramater for loess function.}

\item{verbose}{Boolean.}
}
\value{
SCE object with \code{vg} and \code{vg_info} slots filled for variable genes
and variable gene info. \code{vg_info} is a data frame with log mean, log var, 
fitted log var values, and residuals. \code{vg} is a vector of genes

list with 2 components:
\describe{
  \item{p}{p-value}
  \item{log2fc}{Log base 2 fold change (x over y)}
}
}
\description{
Finds top \code{nf} variable genes by running a loess regression 
on log(variance) and log(mean). Residuals are used to rank 
features, and the top \code{nf} are output.
}
