% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{targets_ids}
\alias{targets_ids}
\title{Return target IDs}
\usage{
targets_ids(x)
}
\arguments{
\item{x}{SCE. SCE object}
}
\value{
Character vector with droplet IDs of called targets
}
\description{
Return target IDs
}
