% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{top100_score_plot}
\alias{top100_score_plot}
\title{Background score density plot}
\usage{
top100_score_plot(x, title = "", return = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{title}{Character}

\item{return}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
This plot is helpful to see the separation of RNA content between low and high count droplets. 
Plot the distribution of background scores for the top 100 count-ranked droplets. 
Plot also contains the background score of the low-count gene probabilities 
used to simulate background droplets.
}
