% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{volcano_de_plot}
\alias{volcano_de_plot}
\title{Volcano plot of differentially expressed genes between nuclear and background droplets}
\usage{
volcano_de_plot(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Volcano plot of differentially expressed genes between nuclear and background droplets
}
