% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{z_table}
\alias{z_table}
\title{Return design matrix of labs vector}
\usage{
z_table(labs)
}
\arguments{
\item{labs}{Vector of labels. Ordered by factor and corresponds to 
columns of returned matrix.}
}
\value{
A droplet by cluster matrix
}
\description{
Return design matrix of labs vector
}
