% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Alpha}
\alias{Alpha}
\title{Get Alpha parameters for clusters}
\usage{
Alpha(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
A droplet by cluster matrix containing the alpha parameters of 
 of the multinomial cluster in each column.
}
\description{
Get Alpha parameters for clusters
}
