% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{assign_clusters}
\alias{assign_clusters}
\title{Assign clusters to droplets from EM}
\usage{
assign_clusters(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
An SCE object.
}
\description{
Assign cluster for each droplet in the test set, along with its 
cluster probability. These can be retrieved using the 
\code{\link{droplet_data}} function.
}
\examples{
\donttest{

sce <- assign_clusters(sce, type="test")
drop_data <- droplet_data(sce)
table(drop_data[,"Cluster"])
summary(drop_data[,"ClusterProb"])

}
}
