% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barcode_rank_plot}
\alias{barcode_rank_plot}
\title{Create a barcode rank plot}
\usage{
barcode_rank_plot(x, title = "", ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{title}{Character}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Create a barcode rank plot
}
