% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barcode_rank_plot}
\alias{barcode_rank_plot}
\title{Create a barcode rank plot}
\usage{
barcode_rank_plot(x, title = "", ret = FALSE)
}
\arguments{
\item{x}{An SCE object}

\item{title}{Title of the plot}

\item{ret}{Logical indicating whether to return a ggplot object}
}
\value{
Nothing. If return=TRUE, then return a ggplot object
}
\description{
Create a barcode rank plot
}
