% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{bin_by_counts}
\alias{bin_by_counts}
\title{Bin by expression counts}
\usage{
bin_by_counts(x, start_bin = 0, end_bin = 150, by_bin = 150)
}
\arguments{
\item{x}{SCE object}

\item{start_bin}{Integer. Read count for starting the bins}

\item{end_bin}{Integer. Read count for ending the bins}

\item{by_bin}{Numeric. Step size, the width of each bin}

\item{Matrix}{with read counts summed across the rows}
}
\description{
Sum read counts for genes across droplets in SCE object. 
Given start, end, and by parameters, sums reads across 
genes in each bin. There are (end - start)/by bins. If 
this is a fraction, the number of bins are rounded down, 
with the same width for each bin.
}
