% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call targets after EM}
\usage{
call_targets(x, pp_thresh = 0.95, min_genes = 200)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{interation}{Integer. DIEM iteration number to use for calliing targets}
}
\value{
SCE object
}
\description{
Call targets from droplets if the log-likelihood membership 
probability is higher than \code{lk_fraction} given during 
initialization.
}
