% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call targets after EM}
\usage{
call_targets(x, p = 0.95)
}
\arguments{
\item{x}{SCE. SCE object}

\item{p}{Numeric. Proportion of likelihood to call a target}
}
\value{
SCE object
}
\description{
Call targets (typically cells or nuclei) from droplets if the 
log-likelihood membership probability is higher than \code{p}.
}
