% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call clean droplets after running EM using percent debris}
\usage{
call_targets(x, thresh_score = 0.5, min_genes = 200, k_init = NULL,
  verbose = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{thresh_score}{The maximum debris score a droplet must have 
to be classified as a cell/nucleus.}

\item{min_genes}{The minimum number of genes a droplet must have 
to be classified as a cell/nucleus.}

\item{k_init}{The k_init run to call droplets from. This must 
specify a single value. If left unspecified, will call droplets only if 
there is one k_init run.}

\item{verbose}{verbosity}
}
\value{
An SCE object.
}
\description{
Call cells or nuclei from an SCE object. Adds meta data to 
the test data slot, which can be retreived
using the function \code{\link{test_data}}. This adds 
column "Call" which takes either "Debris" or "Clean". Droplets
with a debris scores less than \code{thresh_score} and at least 
\code{min_genes} genes detected pass filtering.
}
