% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call targets after EM}
\usage{
call_targets(x, min_count = 200, min_gene = 200, p = 0.95)
}
\arguments{
\item{x}{SCE. SCE object with EM data}

\item{min_count}{Integer. Minimum number of read/UMI counts to call a target}

\item{min_gene}{Integer. Minimum number of genes detected to call a target}

\item{p}{Numeric. Proportion of likelihood to call a target}
}
\value{
SCE object with \code{targets} identified
}
\description{
Call targets, typically cells or nuclei, from droplets if the 
log-likelihood membership probability is higher than \code{p}.
}
