% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{call_targets}
\alias{call_targets}
\title{Call targets after EM}
\usage{
call_targets(x, lk_fraction = 0.95)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{lk_fraction}{Numeric. Select droplets that have a fraction of target likelihood greater than this number.}
}
\value{
SCE object
}
\description{
Call targets from droplets if the log-likelihood membership probability is higher than \code{lk_fraction}.
}
