% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{convert_to_seurat}
\alias{convert_to_seurat}
\title{Convert SCE object to Seurat}
\usage{
convert_to_seurat(x, targets = TRUE, ...)
}
\arguments{
\item{x}{SCE object}

\item{targets}{Boolean. If \code{TRUE}, then output droplets that are specified as targets}

\item{...}{Arguments to \code{\link[Seurat]{CreateSeuratObject}}, such as \code{project} for project name}
}
\value{
A Seurat object
}
\description{
Convert SCE object to Seurat
}
