% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{de_basic}
\alias{de_basic}
\title{Run a basic DE between clusters}
\usage{
de_basic(means, top_n = 20)
}
\arguments{
\item{means}{A gene by cluster sata frame of cluster means, 
of which the columns sum to 1.}

\item{top_n}{Number of top DE genes to return.}
}
\value{
A data frame of DE results
}
\description{
Run a basic DE between clusters
}
