% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{diem}
\alias{diem}
\title{Run DIEM pipeline}
\usage{
diem(sce, top_n = 10000, min_counts = 100, top_thresh = NULL,
  cpm_thresh = 25, eps = 1e-08, max_iter = 100, pp_thresh = 0.95,
  gene_thresh = 200, verbose = TRUE)
}
\arguments{
\item{sce}{SCE. SCE object with raw counts.}

\item{top_n}{Numeric value specifying the  droplets below \code{top_n} (ranked by total counts) 
are fixed to debris.}

\item{min_counts}{A numeric value that specifies that all droplets below this count threshold are 
always fixed to debris.}

\item{top_thresh}{A numeric value specifying any droplets above this count threshold are 
fixed to nuclei. No droplets are fixed to nuclei by default.}

\item{cpm_thresh}{The minimum CPM expression threshold in both groups.}

\item{eps}{Numeric threshold. The EM algorithm converges when the percent change in log likihood is 
less than \code{eps}.}

\item{max_iter}{The maximum number of iterations allowed to run.}

\item{pp_thresh}{Numeric threshold, where clean droplets must have a 
posterior probability of at least \code{pp_thresh}.}

\item{verbose}{Logical indicating verbosity.}

\item{min_genes}{Numeric threshold, where clean droplets must have at least 
\code{min_genes} genes detected.}
}
\value{
SCE object.
}
\description{
Run DIEM pipeline on an SCE object, (see \code{\link{read_10x}} and 
\code{\link{create_SCE}} functions).
}
