% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{diem}
\alias{diem}
\title{Run diem pipeline}
\usage{
diem(sce, n_pcs = 5, n_runs = 10, min_bg_count = 0,
  max_bg_count = 150, min_count = 199, min_gene = 199, simf = 2,
  p = 0.95, seedn = NULL, verbose = TRUE)
}
\arguments{
\item{sce}{SCE. ACE object with raw counts}

\item{n_pcs}{Integer. Number of expression PCs to use for running EM}

\item{n_runs}{Integer. Number of random starts for EM run, with best result returned}

\item{min_bg_count}{Numeric. Minimum number of counts for a droplet to be considered background}

\item{max_bg_count}{Numeric. Maximum number of counts for a droplet to be considered background}

\item{min_count}{Numeric. Minimum number of counts for a droplet to be considered a cell/nucleus}

\item{min_gene}{Numeric. Minimum number of genes for a droplet to be considered a cell/nucleus}

\item{simf}{Numeric. Factor to multiply the number of candidates to get the number of background droplets simulated}

\item{p}{Numeric. Minimum membership probability to call a target}

\item{seedn}{Numeric. Seed for random number generation}

\item{verbose}{Boolean. Print out logging information}
}
\value{
SCE object with droplet IDs labeled as background or nucleus.
}
\description{
Run DIEM on an SCE object, typically returned from \code{\link{read_10x}} function.
DIEM first simulates background droplets by randomly sampling from a multinomial 
distribution. The parameters (gene probabilities) from the distribution are derived 
from droplets with low read counts, as the majority of these are likely to represent 
background RNA contamination. The number of simulated background droplets is given by 
the parameter \code{simf}, which multiplies the number of candidate droplets tested. 
PCA is run on the expression matrix containing the simulated and candidate droplets. 
DIEM then runs semi-supervised EM on the first \code{n_pcs} PCs. In order to call 
droplets as either background or target, the likelihood fraction must be greater than 
\code{p} (0.95 by default). This also calculates the percentage of reads that originate 
from mitochondrial genes, or from MALAT1.
}
