% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{diem}
\alias{diem}
\title{Run diem pipeline}
\usage{
diem(sce, n_pcs = 1, n_runs = 10, min_bg_count = 0,
  max_bg_count = 150, min_tg_count = 199, max_tg_count = Inf,
  min_tg_gene = 199, max_tg_gene = Inf, top_n_tg = NULL,
  top_n_bg = NULL, n_deg = 2000, simf = 1, n_bin = 1,
  shape_mult = 100, p = 0.5, seedn = NULL, verbose = TRUE)
}
\arguments{
\item{sce}{SCE. ACE object with raw counts}

\item{n_pcs}{Integer. Number of expression PCs to use for running EM}

\item{n_runs}{Integer. Number of random starts for EM run, with best result returned}

\item{min_bg_count}{Numeric. Minimum number of counts for a droplet to be considered background}

\item{max_bg_count}{Numeric. Maximum number of counts for a droplet to be considered background}

\item{min_tg_count}{Numeric. Minimum number of counts for a droplet to be considered target}

\item{max_tg_count}{Numeric. Maximum number of counts for a droplet to be considered target}

\item{min_tg_gene}{Numeric. Minimum number of genes detected for a droplet to be considered target}

\item{max_tg_gene}{Numeric. Maximum number of genes detected for a droplet to be considered target}

\item{top_n_tg}{Numeric. Set the top n ranked by counts as the threshold for considering targets}

\item{top_n_bg}{Numeric. Set the top n ranked by counts as the threshold for considering background}

\item{n_deg}{Integer. Number of differentially expressed genes between low/high count droplets to use for calculating PCA}

\item{simf}{Numeric. Factor to multiply the number of candidates to get the number of background droplets simulated}

\item{n_bin}{Numeric. Number of bins to sample from between min_bg_count and max_bg_count}

\item{shape_mult}{Numeric. Number to multiply gene probabilities for sampling from dirichlet}

\item{p}{Numeric. Minimum membership probability to call a target}

\item{seedn}{Numeric. Seed for random number generation}

\item{verbose}{Boolean. Print out logging information}
}
\value{
SCE object with droplet IDs labeled as background or nucleus.
}
\description{
Run DIEM on an SCE object, typically returned from \code{\link{read_10x}} function.
DIEM first simulates background droplets by randomly sampling from a multinomial 
distribution. The parameters (gene probabilities) from the distribution are derived 
from droplets with low read counts, as the majority of these are likely to represent 
background RNA contamination. The number of simulated background droplets is given by 
the parameter \code{simf}, which multiplies the number of candidate droplets tested. 
PCA is run on the expression matrix containing the simulated and candidate droplets. 
DIEM then runs semi-supervised EM on the first \code{n_pcs} PCs. In order to call 
droplets as either background or target, the likelihood fraction must be greater than 
\code{p} (0.95 by default). This also calculates the percentage of reads that originate 
from mitochondrial genes, or from MALAT1.
}
