% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diem.R
\name{diem}
\alias{diem}
\title{Run diem pipeline}
\usage{
diem(sce, log_base_em = 10, top_n = NULL, min_counts = NULL,
  log_base_label = 5, log_base_de = 100, de_cutpoint = NULL,
  cpm_thresh = 3, log2fc_thresh = 0.25, scale_factor = 1,
  logt = TRUE, min_iter = 5, max_iter = 1000, eps = 1e-10,
  n_runs = 10, lk_fraction = 0.95, seedn = NULL, verbose = TRUE)
}
\arguments{
\item{sce}{SCE. SCE object with raw counts.}

\item{log_base_em}{Numeric. Take droplets with counts above 2 logs below the max droplet count using this base
for EM classification}

\item{top_n}{Numeric. Only run EM to classify droplets if they are in the 
top number of droplets ranked by total counts.}

\item{min_counts}{Integer. Only run EM to classify droplets with at least this number of total counts.}

\item{log_base_label}{Numeric. Fix the labels as nuclei for droplets with counts above 1 log below 
the max droplet count using this base.}

\item{de_cutpoint}{Numeric. Manually set the DE cutpoint. If not NULL, it is calculated from the data.}

\item{log2fc_thresh}{Numeric. Test only genes that have an absolute log2 fold-change above this threshold.}

\item{scale_factor}{Numeric. A scaling factor to multiply values after division by total count.}

\item{min_iter}{Integer. Minimum number of iterations in EM.}

\item{max_iter}{Integer. Maximum number of iterations in EM.}

\item{eps}{Numeric. Epsilon of log-likilhood change to call convergence in EM.}

\item{n_runs}{Integer. Number of EM initializations to run, with best result returned.}

\item{lk_fraction}{Numeric. Select droplets that have a fraction of target likelihood greater than this number.}

\item{seedn}{Numeric. Seed for random number generation.}

\item{verbose}{Boolean. Print out logging information.}

\item{de_p_thresh}{Numeric. Adjusted p-value threshold.}

\item{de_correct}{Character. p-value correction method.}
}
\value{
SCE object.
}
\description{
Run DIEM pipeline on an SCE object, typically returned from \code{\link{read_10x}} function.
}
