% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{divide_by_colsum}
\alias{divide_by_colsum}
\title{Divide elements of a column by the column's sum in a sparse matrix}
\usage{
divide_by_colsum(x)
}
\arguments{
\item{x}{A Sparse Matrix.}
}
\value{
The Sparse Matrix x with columns summing to 1.
}
\description{
Divide elements of a column by the column's sum in a sparse matrix
}
