% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{dmultinom_sparse}
\alias{dmultinom_sparse}
\title{Get log multinomial density of columns in a sparse matrix.}
\usage{
dmultinom_sparse(X, prob)
}
\arguments{
\item{X}{A sparseMatrix that is a sample by feature matrix of counts.}

\item{prob}{Probability parameter of multinomial. Must be same size as 
number of columns in X.}
}
\description{
Get log multinomial density of columns in a sparse matrix.
}
