% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{Run EM}
\usage{
em(counts, params, labs, eps = 1e-04, max_iter = 100, verbose = TRUE)
}
\arguments{
\item{counts}{A gene by droplet matrix}

\item{params}{A list containing
\describe{
     \item{Alpha}{ A gene by cluster matrix with parameters of 
         Dirichlet-multinomial.}
     \item{Pi}{ A vector of mixing coefficients. Must sum to 1 
         and be between 0 and 1.}
     }}

\item{labs}{Vector of labels for droplets. A label of 0 means unknwon 
while 1 means background.}

\item{eps}{The threshold of change in the parameter to stop the 
EM. The parameter checked is the average change in posterior 
probabilities Z.}

\item{max_iter}{Maximum number of iterations.}

\item{verbose}{Verbosity.}
}
\description{
Run EM
}
