% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{EM function
Counts is sample x feature}
\usage{
em(counts, k, mn_params, max_iter = 1000, eps = 1e-08, psc = 1e-04,
  labels = NULL, verbose = TRUE)
}
\description{
EM function
Counts is sample x feature
}
