% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{estimate_dbr_score}
\alias{estimate_dbr_score}
\title{Estimate debris score per droplet}
\usage{
estimate_dbr_score(x, thresh_score = 0.5, thresh_genes = 200,
  thresh_p = 0.05, p_method = "fdr", dbr_fixed = NULL,
  verbose = TRUE, k_init = NULL)
}
\arguments{
\item{x}{An SCE object.}

\item{thresh_score}{Threshold for cluster mean debris score. Clusters 
with a mean above this value are considered debris clusters.}

\item{thresh_genes}{Threshold for cluster mean number of genes 
detected. Clusters with a mean number of genes less than this value 
are set to debris clusters.}

\item{thresh_p}{P-value threshold for calculating differential 
expression between droplets in the debris and non-debris clusters.}

\item{p_method}{Method for p-value correction using 
\code{\link[stats]{p.adjust}}.
expression between droplets in the debris and non-debris clusters.}

\item{dbr_fixed}{Specifies additional clusters to set as debris. Cluster 
1 is always included. Use this only if you know what you are doing.}

\item{verbose}{Verbosity, indicating whether to show a progress
bar.}

\item{k_init}{Run EM on the \code{k_init} initialization(s). 
If NULL (default), run on all \code{k_init} initializations.}
}
\value{
An SCE object
}
\description{
Estimate debris score per droplet
}
