% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{fill_dropl_info}
\alias{fill_dropl_info}
\title{Add calls, PP, and LLK to dropl_info.}
\usage{
fill_dropl_info(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
An SCE object with call info, PP, and log-likilhood added to dropl_info.
}
\description{
Add calls, PP, and LLK to dropl_info.
}
