% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter out lowly expressed features}
\usage{
filter_genes(x, cpm_thresh = 10)
}
\arguments{
\item{x}{An SCE object.}

\item{cpm_thresh}{The minimum CPM expression threshold in both groups.}
}
\value{
An SCE object
}
\description{
This function removes genes that are lowly expressed in 
either the nuclear or background group. Read/UMI counts are summed across 
droplets in the nuclear-enriched and background-enriched groups. The 
background-enriched droplets are those with labels fixed to debris, 
(see \code{\link{set_test_set}}). Counts per million mapped reads (CPM) 
are calculated for the two groups, and only genes with a CPM of at least 
\code{cpm_thresh} in both groups are kept. This function ensures that 
the likelihood of the multinomial mixture does not collapse to 0, since 
a gene with 0 counts has a likelihood of 0. This creates a logical 
vector in the gene_info slot that indicates whether a gene passes the 
threshold or not.  #'
}
