% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fix_tails}
\alias{fix_tails}
\title{Fix labels of lower and upper count droplets}
\usage{
fix_tails(x, pct = 0)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{pct}{Numeric. Pct of tails to fix, fraction from 0 to 1.}
}
\value{
An SCE object
}
\description{
Fix labels of the upper and lower tails droplets to signal and background, respectively.
The top and bottom \code{pct} are then fixed during EM.
}
