% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_clean_ids}
\alias{get_clean_ids}
\title{Return IDs of clean droplets}
\usage{
get_clean_ids(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
A character vector with the called droplet IDs.
}
\description{
Return the barcodes that pass DIEM filtering. The function 
\code{\link{call_targets}} must be run beforehand.
}
