% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{get_de}
\alias{get_de}
\title{Get DE genes between high and low read count droplets using a t-test}
\usage{
get_de(x, log2fc_thresh = 0.5, de_p_thresh = 0.05,
  de_correct = "bonferroni", verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{log2fc_thresh}{Numeric. Test only genes that have this log2 fold-change threshold between low and high
droplets will be considered as differentially expressed. Fold change is for both comparison low vs high and high vs low.}

\item{de_p_thresh}{Numeric. Adjusted p-value threshold.}

\item{de_correct}{Character. p-value correction method.}

\item{verbose}{Boolean.}
}
\value{
SCE object.
}
\description{
Find differentially expressed genes between low and high read count droplets. Groups droplets into
low or high if they have less than or equal to or greater than \code{de_cutpoint}, respectively. Must run
set \code{de_cutpoint} by running \code{\link{set_de_cutpoint}} before.A t-test is run between the 2 groups 
for each gene, calculating log fold change and p-value. Statistically significant genes are determined by adjusting 
the p-value using the \code{de_correct}. Can be any method used in \code{\link[stats]{p.adjust}}, typically either 
"bonferroni" (default) or "fdr."
}
