% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{get_de_genes}
\alias{get_de_genes}
\title{Get DE genes between high and low read count droplets using a t-test}
\usage{
get_de_genes(x, de_cutpoint = NULL, log2fc_thresh = 0,
  de_p_thresh = 0.05, de_correct = "fdr", verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{de_cutpoint}{Numeric. Manually set the DE cutpoint. If not NULL, it is calculated from the data.}

\item{log2fc_thresh}{Numeric. Test only genes that have an absolute log2 fold-change above this threshold.}

\item{de_p_thresh}{Numeric. Adjusted p-value threshold.}

\item{de_correct}{Character. p-value correction method.}

\item{verbose}{Boolean.}
}
\value{
SCE object.
}
\description{
Find differentially expressed genes between low and high read count droplets. Groups droplets into
low or high count groups. The count threshold for separating low and high is determined by taking the 
log10 of the maximum droplet count, subtracting 1, and calculating 10 to the power of this number. In other 
words, the thresold is found by dividing the max count by 10. 
Alternatively, the you can set the cutpoint threshold manually with \code{de_cutpoint}. A t-test is run between the 2 groups 
for each gene, calculating log fold change and p-value. Statistically significant genes are determined by adjusting 
the p-value using the \code{de_correct}. Can be any method used in \code{\link[stats]{p.adjust}}, typically either 
"bonferroni" or "fdr" (default).
}
