% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_de_genes}
\alias{get_de_genes}
\title{Get DE genes between high and low read count droplets}
\usage{
get_de_genes(x, n_genes = 500, sd = NULL, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{n_genes}{Numeric. Number of genes to output as differentially expressed between background and candidate}

\item{verbose}{Boolean.}
}
\value{
SCE object with \code{gene_info[,"diff_prop"]} and \code{deg} slots filled. 
The \code{diff_prop} column contains the background - candidate difference in proportions.
The \code{deg} slot contains the names of the genes output as DE.
}
\description{
Find differentially expressed genes between low and high read count droplets. Sums
gene expression across all cells within the limits specified in the SCE object, 
divides by the total number of reads in each, and calculates the difference between 
the 2. The top \code{n_genes} genes sorted by absolute difference in proportion are 
designated as DE.
}
