% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{get_dist}
\alias{get_dist}
\title{Get cluster distances to background disrtibution}
\usage{
get_dist(x, k_init = NULL, verbose = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{k_init}{An integer specifying the k_init run(s) to remove clusters 
from. The default is to remove from all k_init values.}

\item{verbose}{Verbosity.}
}
\value{
An SCE object
}
\description{
Get the likelihood-based distance of the clusters to the background 
distribution. The distances take values between 0 and 1, where a 
larger value for a cluster indicates it is further away and thus more 
likely to reprsent a cell type. The distance values can plotted against 
average number of UMIs in a cluster using \code{\link{plot_dist}} to 
help decide which threshold to use. It is recommended to use a threshold 
so that improperly initialized clusters can be removed. The distances 
can be retrieved by calling the \code{\link{distances}} function.
}
