% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_gene_pct}
\alias{get_gene_pct}
\title{Get percent of reads aligning to given genes.}
\usage{
get_gene_pct(x, genes, name)
}
\arguments{
\item{x}{An SCE object.}

\item{genes}{Genes to calculate percentage of in counts.}

\item{name}{Column name to place in dropl_info.}
}
\value{
An SCE object.
}
\description{
Get percent of reads aligning to given genes.
}
\examples{
mm_seur <- get_gene_pct(x=mb_sce, genes="Malat1", name="pct.malat1")
mt_genes <- grep(pattern="^mt-", x=rownames(mb_sce), ignore.case=TRUE)
mm_seur <- get_gene_pct(x=mb_sce, genes=mt_genes, name="pct.mt")
}
