% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_llk}
\alias{get_llk}
\title{log likelihood of multivariate normal under 1 or more latent clusters.}
\usage{
get_llk(x, mu, sgma, tau)
}
\arguments{
\item{x}{matrix or data frame. Observations with which to calculate likelihood under parameters.}

\item{mu}{list of numeric vectors. Averages for each of the k clusters.}

\item{sgma}{list of numeric vectors. Covariance for each of the k clusters.}

\item{tau}{numeric vector. Mixing parameters for each of the k clusters.}
}
\value{
numeric value of the observed log likelihood under k clusters.
}
\description{
Calculates log likelihood of multivariate distribution given data and parameters. 
For each of k clusters, calculate log likelihood, multiply by its mixing parameter tau, and 
sum across clusters.
}
