% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_pcs}
\alias{get_pcs}
\title{Get PCs from expression data}
\usage{
get_pcs(x, n_pcs = 15, center = TRUE, scale. = TRUE)
}
\arguments{
\item{x}{SCE. An SCE object}

\item{n_pcs}{Integer. Number of PCs to calculate}

\item{center}{Boolean. Whether to center variables}

\item{scale.}{Boolean. Whether to scale variables to have unit variance}
}
\value{
An SCE object with \code{\link[irlba]{prcomp_irlba}} output in 
the \code{pcs} slot
}
\description{
Get PCs from expression data
}
