% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_pcs}
\alias{get_pcs}
\title{Get PCs from expression data}
\usage{
get_pcs(x, n_pcs = 30, genes = NULL, verbose = FALSE)
}
\arguments{
\item{x}{SCE. An SCE object.}

\item{n_pcs}{Integer. Number of PCs to calculate.}

\item{genes}{Vector. Use these genes specified.}

\item{verbose}{Boolean.}
}
\value{
An SCE object with \code{\link[irlba]{prcomp_irlba}} output in 
the \code{pcs} slot
}
\description{
Runs irlba implementation of prcomp. Takes normalized expression data in 
\code{x@sim@norm}, subsets using genes in the \code{x@de@deg}.
The first \code{n_pcs} are calculated.
The return object for \code{prcomp_irlba} is placed in \code{x@pcs}. The 
principal component scores for the droplets are thus stored in \code{x@pcs$x}.
}
