% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_pcs}
\alias{get_pcs}
\title{Get PCs from normalized count data}
\usage{
get_pcs(x, n_pcs = 10, use_var = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{n_pcs}{Number of PCs to output.}
}
\value{
An SCE object.
}
\description{
Get PCs from normalized count data
}
