% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{get_pi}
\alias{get_pi}
\title{Get mixing coefficients of Mixture model from posterior probabilities Z}
\usage{
get_pi(Z)
}
\arguments{
\item{Z}{A droplet by cluster matrix of the posterior probability a 
droplet belongs to cluster k. Rows must sum to 1 and entries must 
be between 0 and 1.}
}
\value{
A cluster-length numeric vector that sums to 1 and entries 
 between 0 and 1.
}
\description{
Get mixing coefficients of Mixture model from posterior probabilities Z
}
