% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{get_removed_ids}
\alias{get_removed_ids}
\title{Return IDs of removed droplets}
\usage{
get_removed_ids(x)
}
\arguments{
\item{x}{An SCE object.}
}
\value{
A character vector with the called droplet IDs.
}
\description{
Return the droplet IDs in the test set that have been removed 
by DIEM. 
The function \code{\link{call_targets}} must be run beforehand.
}
