% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{init}
\alias{init}
\title{Initialize parameters}
\usage{
init(x, k_init = 20, use = "kmeans", droplets.use = NULL,
  n_sample = NULL, nstart_init = 30, min_size_init = 10,
  fixed = NULL, model = "mltn", psc = 1e-10, seedn = 1,
  threads = threads, verbose = TRUE, ...)
}
\arguments{
\item{x}{An SCE object.}

\item{k_init}{The number of clusters to initialize, not including 
fixed background clusters.}

\item{use}{The method to use for clustering, one of either 
"kmeans" or "hclust."}

\item{droplets.use}{Specify droplet IDs to use for initializing 
parameters.}

\item{n_sample}{The number of droplets to sample from for initializing
parameters.}

\item{nstart_init}{The number of starts to use in k-means for 
for the initialization.}

\item{min_size_init}{The minimum number of droplets that must belong 
to an initialized cluster.}

\item{fixed}{A named integer vector that specifies which droplets to 
fix to which clusters. If \code{NULL}, then sets background droplets 
to 1.}

\item{model}{The mixture model to assume. Can be either "DM" for 
a Dirichlet-multinomial or "mltn" for a multinomial.}

\item{psc}{Pseudocount to add to estimation of gene probabilities.}

\item{seedn}{Random seed.}

\item{threads}{Number of threads.}

\item{verbose}{Verbosity.}

\item{...}{Additional parameters to pass to hclust or kmeans.}
}
\value{
An SCE object
}
\description{
Initialize the parameters of the mixture model using 
k-means or hierarchical clustering on the PCs.
}
\examples{
\donttest{

# Initialize parameters with default values and multiple threads
sce <- init(sce, threads = 8)

# Specify initial k of 10 cell type clusters
sce <- init(sce, k_init = 10, threads = 8)

# Initialize parameters for Dirichlet-multinomial
sce <- init(sce, k_init = 10, model = "DM", threads = 8)

# Set seedn to NULL for random starts
sce <- init(sce, k_init = 10, seedn = NULL, threads = 8)
sce <- init(sce, k_init = 10, seedn = NULL, threads = 8)
sce <- init(sce, k_init = 10, seedn = NULL, threads = 8)

# Specify initial k of 30 cell type clusters and 
# Only allow clusters with at least 30 droplets
sce <- init(sce, k_init = 30, min_size_init = 30, threads = 8)
}

}
