% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{init}
\alias{init}
\title{Initialize clusters using hierarchical clustering
k is a positive integer or vector of positive integers}
\usage{
init(x, n_pcs = 30, k_init = 30, iter.max = 15, nstart = 30,
  min_size = 10, fltr = 0.1, seedn = 1, psc = 1e-10,
  verbose = TRUE)
}
\description{
Initialize clusters using hierarchical clustering
k is a positive integer or vector of positive integers
}
