% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{init_best}
\alias{init_best}
\title{Initialize best from random starts}
\usage{
init_best(X, labs = NULL, K, hyperp, top_n = 1000, init_start = 5,
  init_iter = 3, verbose = TRUE)
}
\description{
Initialize best from random starts
}
