% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{init_param}
\alias{init_param}
\title{Initialize EM parameters with very droplet assigned to a group}
\usage{
init_param(counts, groups, psc = 1e-04)
}
\value{
List with mu0 a p x 2 matrix and mc0 a length 2 numeric
}
\description{
Initialize EM parameters with very droplet assigned to a group
}
