% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{init_tails}
\alias{init_tails}
\title{Initialize EM parameters for k=2 low and high groups}
\usage{
init_tails(x, counts, pct = 0.1)
}
\arguments{
\item{x}{SCE. SCE object with pi calculated}

\item{pct}{Numeric. Percent of droplets at each tail to fix. A number from 0 to 1.}
}
\value{
List with mu0 a p x 2 matrix and tau0 a length 2 numeric
}
\description{
Initialize the means of the two multinomial groups.
}
