% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{initialize_clusters}
\alias{initialize_clusters}
\title{Initialize clusters for unlabeled data}
\usage{
initialize_clusters(x, bf_thresh = 10, min_size = 30, gammas = c(10,
  1e+07), tol_opt = 100)
}
\value{
Numeric vector of memberships
}
\description{
Initialize clusters for unlabeled data
}
