% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{initialize_clusters}
\alias{initialize_clusters}
\title{Initialize clusters for unlabeled data}
\usage{
initialize_clusters(x, cluster_n = 500, order_by = "gene", nn = 30,
  min_size = 20, verbose = TRUE)
}
\value{
Numeric vector of memberships
}
\description{
Initialize clusters for unlabeled data
}
