% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clust.R
\name{initialize_clusters}
\alias{initialize_clusters}
\title{Initialize clusters for unlabeled data}
\usage{
initialize_clusters(x, nn = 30, method = "louvain", bf_thresh = 10,
  gammas = c(10, 1e+07), tol_opt = 100)
}
\value{
Numeric vector of memberships
}
\description{
Initialize clusters for unlabeled data
}
