% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{initialize_clusters}
\alias{initialize_clusters}
\title{Initialize clustering}
\usage{
initialize_clusters(x, K = 30, n_start = 10, km_iter = 3,
  verbose = FALSE)
}
\arguments{
\item{x}{An SCE object.}

\item{K}{Number of clusters to return from k-means. This gives 
the Maximum number of clusters}

\item{verbose}{verbosity.}
}
\value{
An SCE object
}
\description{
Instead of randomly initializing the EM, cell types are estimated from 
droplets that are expected to contain cells/nuclei. The initialization 
is done with droplets in the cluster set. The data is then normalized 
by first calculating the variable genes. A loess regression line 
is fit between the log counts and log variance, and the only top genes 
ranked by residual are used to initialize the clusters. The number of 
genes is specified with \code{n_var}. Optionally, one can use all genes 
by setting \code{use_var} to FALSE. The span of the loess regression line 
is given by \code{lss} (default is 0.3).
The data is normalized by dividing counts by the total counts per droplet. 
Then, the counts are multiplied by a scaling factor, given by 
\code{sf} (the median of total counts by default). Finally, the data is 
log transformed after adding a constant value of 1.
After normalization, the k-nearest neighbors are identified in the 
cluster set. The number of nearest neighbors is specified by 
\code{nn}. Clusters are identified from the KNN graph 
using the Louvain algorithm. Finally, only clusters with at least 
\code{min_size} (20 by default) droplets are considered cell types.
}
