% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{llk_fraction_plot}
\alias{llk_fraction_plot}
\title{Distribution of likelihood fraction}
\usage{
llk_fraction_plot(x, ret = FALSE)
}
\arguments{
\item{x}{SCE. SCE object}

\item{ret}{Boolean. Return a ggplot object}
}
\value{
Nothing, unless return=TRUE then a ggplot
}
\description{
Distribution of likelihood fraction
}
