% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb_small}
\alias{mb_small}
\title{Single-nucleus RNA-seq of mouse brain}
\format{An SCE object with 1,001 rows (genes) and 2,451 droplets (columns)}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/nuclei_2k}
}
\usage{
mb_small
}
\description{
A dataset containing a mostly random subset of the publicly available 
10X single-nucleus RNA-seq data set of freshly dissected 
mouse brain tissue. This subset contains 1,001 genes and 
2,451 droplets, including all mitochondrial genes and Malat1. 
This dataset is distributed under the creative commons license 
\url{https://creativecommons.org/licenses/by/4.0/}.
}
\keyword{datasets}
