% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_counts}
\alias{norm_counts}
\title{Normalize counts of a sparse matrix}
\usage{
norm_counts(counts, scale_factor = 10000, logt = TRUE)
}
\arguments{
\item{counts}{A Sparse Matrix.}

\item{scale_factor}{A numeric scaling factor to 
multiply values after dividing counts matrix by total counts.}

\item{logt}{A logical indicating whether to log transform after 
normalizing for columns to sum to 1 and multiplying by scaling factor.}
}
\value{
Sparse Matrix
}
\description{
Normalize counts of a sparse matrix
}
