% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{normalize}
\alias{normalize}
\title{Normalize raw count data}
\usage{
normalize(x, logt = TRUE, sf = NULL, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE}

\item{logt}{Boolean. Log transform}

\item{sf}{Numeric. A scaling factor to multiply values after division by total count}

\item{verbose}{Boolean.}
}
\value{
An SCE object with normalized expression data
}
\description{
Raw expression counts are divided by total counts per droplet, 
multiplied by scaling factor, and (optionally) log1p transformed (default). 
This normalizes the counts in 
\code{raw} slot and places them in the @norm slot. The scaling factor 
defaults to the median total read count across all droplets.
}
