% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normalize}
\alias{normalize}
\title{Normalize simulated and candidate raw count data}
\usage{
normalize(x, scale_factor = 10000)
}
\arguments{
\item{x}{SCE.}

\item{scale_factor}{Numeric. A scaling factor to multiply values after division by total count.}
}
\value{
SCE object
}
\description{
Counts are normalized by dividing by total counts per droplet, and log1p-transformed. 
Optionally, counts can be multiplied by a scaling factor \code{scale_factor}.
}
