% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normalize}
\alias{normalize}
\title{Normalize raw counts for droplets being classified.}
\usage{
normalize(x, genes = NULL, scale_factor = 1, logt = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{SCE.}

\item{scale_factor}{Numeric. A scaling factor to multiply values after division by total count.}

\item{logt}{Boolean. Whether to log1p transform expression values.}
}
\value{
SCE object
}
\description{
Counts are normalized by dividing by total counts per droplet. Optionally, droplets can be normalized 
by log1p transformation, with a counts being multiplied by a scaling factor \code{scale_factor}.
Additionally, gene counts can be scaled to mean 0 and variance 1 (default). This is recommended so that 
highly expressed genes do not drive the calculation of pi. Note that, when scaling, genes with 0 mean 
and zero variance will be removed automatically.
}
