% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize raw counts.}
\usage{
normalize_data(x, droplets.use = NULL, genes.use = NULL,
  use_var = FALSE, n_var = 2000, lss = 0.3, logt = TRUE)
}
\arguments{
\item{x}{SCE.}

\item{genes.use}{A character vector to subset count data to. Normalization will only be run for these genes.}

\item{logt}{Logical indicating whether to log1p transform expression values.
Default is TRUE}

\item{scale_factor}{A numeric scaling factor to multiply counts after division by column sum
Default is 1.}
}
\value{
SCE object
}
\description{
Counts are normalized by dividing by total counts per droplet. Note that this only normalizes 
droplets that are candidates in the test set. Optionally, droplets can be normalized 
by log1p transformation, with a counts being multiplied by a scaling factor \code{scale_factor}.
}
