% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_10x}
\alias{read_10x}
\title{Read 10X output}
\usage{
read_10x(path)
}
\arguments{
\item{path}{Character. File path prefix to 10X output.}
}
\value{
expr Sparse matrix. Expression counts from 10X cell ranger
 with cells in the columns and genes in the rows.
}
\description{
Read output from 10X into a sparse matrix. Given path, reads 
files path/matrix.MM path/genes.tsv, and path/barcodes.tsv
}
