% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{run_em}
\alias{run_em}
\title{Run EM to estimate parameters of mixture model}
\usage{
run_em(x, A = NULL, P = NULL, Z = NULL, model = "mltn",
  alpha_prior = 0, pi_prior = 0, max_iter = 1000, eps = 1e-04,
  psc = 1e-10, threads = 1, verbose = TRUE)
}
\arguments{
\item{x}{An SCE object.}

\item{A}{Initial alpha parameter.}

\item{P}{Initial pi parameter.}

\item{Z}{Initial latent variable Z values.}

\item{model}{The mixture model to assume. Can be either "DM" for 
a Dirichlet-multinomial or "mltn" for a multinomial.}

\item{alpha_prior}{Add a non-informative prior by adding a count of
\code{alpha_prior} to all genes in each cluster. Only valid for 
the multinomial model.}

\item{pi_prior}{Add a non-informative prior by adding a count of 
\code{pi_prior} to the each cluster's membership.}

\item{max_iter}{Maximum number of iterations for the EM estimation 
of the mixture model.}

\item{eps}{The delta threshold for when to call convergence for 
the EM estimation of the mixture model. The EM 
stops when delta falls below this value. We define delta as the 
average change in posterior probability. By default this is set to 
1e4, so that the EM converges when less than 1 in 10,000 labels 
change on average.}

\item{psc}{Pseudocount to add to avoid collapsing likelihood to 0.}

\item{threads}{Number of threads for parallel execution. Default is 1.}

\item{verbose}{Verbosity.}
}
\value{
An SCE object.
}
\description{
Estimate the parameters of the mixture model, 
and estimate the posterior probability a droplet belongs to each 
of the clusters.
}
