% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{run_em}
\alias{run_em}
\title{Run EM
Run EM for each initialized K, storing output in SCE object.}
\usage{
run_em(x, k_init = NULL, eps = 1e-04, fltr = 0.1, max_iter = 100,
  verbose = TRUE)
}
\description{
Run EM
Run EM for each initialized K, storing output in SCE object.
}
