% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{run_em}
\alias{run_em}
\title{Run EM on pi features}
\usage{
run_em(x, eps = 1e-08, max_iter = 1000, labels = NULL,
  verbose = TRUE)
}
\arguments{
\item{x}{SCE object}
}
\value{
SCE object with EM output in the \code{emo} slot. See \code{\link{run_mv_em_diag}} for details
}
\description{
Run expectation maximization (EM) using diagonal covariance on pi matrix in an SCE object with k=2 groups. 
EM is run \code{n_runs} times with random initializations, and parameters with 
the best log likelihood are output. Within EM, iterate a minimum of \code{min_iter} times and 
a maximum of \code{max_iter} times. The \code{eps} value gives the threshold of percentage change in 
log likelihood until convergence is reached.
}
