% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debris_id.R
\name{run_em_pcs}
\alias{run_em_pcs}
\title{Run EM on PCs}
\usage{
run_em_pcs(x, n_pcs = NULL, min_iter = 5, max_iter = 1000,
  eps = 1e-10, n_runs = 5, seedn = NULL, verbose = TRUE)
}
\arguments{
\item{x}{SCE. SCE object with PCs and labels}

\item{n_pcs}{Integer. Number of expression PCs to use for running EM}

\item{min_iter}{Integer. Minimum number of iterations}

\item{max_iter}{Integer. Maximum number of iterations}

\item{eps}{Numeric. Epsilon of log-likilhood change to call convergence}

\item{n_runs}{Integer. Number of EM runs, with best result returned}

\item{seedn}{Numeric. Seed for random number generation}

\item{verbose}{Boolean. Print out logging information}
}
\value{
SCE object with EM output in the \code{emo} slot. See \code{\link{run_mv_em_diag}} for details
}
\description{
Run expectation maximization (EM) using diagonal covariance on expression PCs in an SCE object with k=2 groups. 
EM is run \code{n_runs} times with random initializations, and the EM output parameters with 
the best log likelihood is output. Within EM, iterate a minimum of \code{min_iter} times and 
a maximum of \code{max_iter} times. The \code{eps} value gives the threshold of percentage change in 
log likelihood until covnergence is called. \code{seedn} gives the seed for random number generation, 
which can be used to reproduce results.
}
