% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{run_mv_em_diag}
\alias{run_mv_em_diag}
\title{Run multivariate EM restricted to diagonal covariance}
\usage{
run_mv_em_diag(x, k = 2, min_iter = 5, max_iter = 1000,
  labels = NULL, mu0 = NULL, sgma0 = NULL, tau0 = NULL,
  eps = 1e-10, seedn = NULL, verbose = TRUE, trace = FALSE)
}
\arguments{
\item{x}{matrix. Matrix with samples on the rows and features on the columns.}

\item{k}{Numeric. Number of clusters. Must be pre-specified.}

\item{min_iter}{Numeric. Minimum number of iterations before calling convergence.}

\item{max_iter}{Numeric. Maximum number of iterations.}

\item{labels}{Numeric vector. Fixed label IDs for semi-supervised clustering. 0 is unknown (to be estimated).}

\item{mu0}{Numeric. Starting value for mu. If set, all mu0, sgma0, and tau0 must be set.}

\item{sgma0}{Numeric. Starting value for sigma (variance). If set, all mu0, sgma0, and tau0 must be set.}

\item{tau0}{Numeric. Starting value for tau. If set, all mu0, sgma0, and tau0 must be set.}

\item{eps}{Numeric. Threshold of change in log likelihood to stop iterations.}

\item{seedn}{Numeric. Numeric value specifying seed for random initialization.}

\item{verbose}{Boolean. Print logging information.}

\item{trace}{Boolean. Print out parameter and likelihod values during iterations.}
}
\value{
EMO object
}
\description{
Run multivariate un-, semi-, or fully supervised EM.
}
