% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clust.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(x, n_pcs = 30)
}
\arguments{
\item{x}{An SCE object.}

\item{n_pcs}{Number of PCs to calculate.}
}
\value{
An SCE object with PCs
}
\description{
Get PCs from normalized counts.
}
