% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SCE-class}
\alias{SCE-class}
\alias{SCE}
\title{SCE}
\description{
The SCE class
}
\details{
This class is used to store the expression data from a 
single-cell RNA-seq experiment. It is the class that is used by 
DIEM. The results from filtering are stored in the \code{droplet_data} 
slot, and the converged EM parameters are stored in a list in the 
\code{emo} slot.

Single Cell Expression object
}
