% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{set_cluster_set}
\alias{set_cluster_set}
\title{Set droplets for cluster initialization}
\usage{
set_cluster_set(x, cluster_n = 1000, order_by = "gene",
  verbose = FALSE)
}
\arguments{
\item{x}{An SCE object.}

\item{cluster_n}{Numeric value specifying the number of droplets to use 
in the cluster set. The top \code{cluster_n} droplets, ranked by 
count or gene, are included in the cluster set.}

\item{order_by}{Whether to order the droplets by total number of total 
counts or total number of genes detected.}

\item{verbose}{verbosity}
}
\value{
An SCE object.
}
\description{
This function sets the droplets that will be used for the 
initialization. The cell types from clustering these droplets are used to
to intialize the parameters of the multinomal for the EM. The top 
\code{cluster_n} droplets ranked by either gene or count are fixed as 
to the debris group.
}
