% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_test_set}
\alias{set_test_set}
\title{Subset droplets for EM testing}
\usage{
set_test_set(x, top_n = 10000, min_counts = 30, verbose = FALSE)
}
\arguments{
\item{x}{SCE. SCE object.}

\item{top_n}{Numeric. Only the top \code{top_n} droplets ranked by total counts are classified.}

\item{min_counts}{Integer. Only droplets with at least this number of total counts are classified. 
Can override \code{top_n} if \code{top_n} includes droplets with less than \code{min_counts}.}
}
\value{
An SCE object.
}
\description{
Sets the threshold for testing droplets and subsets counts matrix. Includes the top \code{top_n} 
droplets for testing (including any ties), and removing any droplets with total counts 
less than \code{min_counts}. The count matrix is subset to these droplets for EM testing
}
