% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_test_set}
\alias{set_test_set}
\title{Set droplets for EM testing}
\usage{
set_test_set(x, top_n = 10000, min_counts = 150, min_genes = 150)
}
\arguments{
\item{x}{An SCE object.}

\item{top_n}{Numeric value specifying the  droplets below \code{top_n} (ranked by total counts) 
are fixed to debris. Set to NULL to ignore.}

\item{min_counts}{A numeric value that specifies that all droplets below this count threshold are 
always fixed to debris.}

\item{top_thresh}{A numeric value specifying any droplets above this count threshold are 
fixed to nuclei. No droplets are fixed to nuclei by default.}
}
\value{
An SCE object.
}
\description{
Sets the count threshold \emph{T} for testing droplets. Droplets with counts below \emph{T} 
are fixed as debris, while droplets greater than or equal to \emph{T} are set as unlabeled. The 
threshold \emph{T} is found by ranking droplets in order of decreasing size (total read/UMI counts) 
and taking the read/UMI count size of the \code{top_n} barcode. If this ranked count is less than 
\code{min_counts}, then set \emph{T} to \code{min_counts} instead. In addition to fixing the labels 
of debris droplets, it is possible to fix droplets as clean. Droplets with counts above \code{top_thresh} 
are fixed to nuclei. By default, no droplets are fixed as clean and this is not recommended.
}
