% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{subset_n_dropls}
\alias{subset_n_dropls}
\title{Subset to top n droplets in SCE object}
\usage{
subset_n_dropls(x, n, slot_name = "raw")
}
\arguments{
\item{x}{SCE object. Must contain data in raw slot}

\item{n}{Integer. Number of droplets to subset}

\item{slot_name}{Character. Slot of SCE to subset, either raw or norm, or both}
}
\value{
x SCE object subsetted
}
\description{
Subset to top n droplets in SCE object
}
