% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{summarize_clusters}
\alias{summarize_clusters}
\title{Output summary stats from a DIEM run}
\usage{
summarize_clusters(x, top_n = 20)
}
\arguments{
\item{x}{An SCE object.}

\item{top_n}{Number of top DE genes to return in column.}
}
\value{
A data frame
}
\description{
Return summary statistics from a DIEM run to evaluate clustering 
and classification. The function calculates helpful statistics to 
evaluate each cluster, including the debris and putatitive cell 
types. This returns a data frame with the following 
summary stats per cluster:
\describe{
\item{Type}{Whether the cluster was considered as debris for 
 calculating the debris score.}
\item{Cluster}{Cluster for which summary stats are shown.}
\item{n_droplets}{Number of droplets.}
\item{avg_n_genes}{Average number of genes detected.}
\item{avg_n_counts}{Average number of counts.}
\item{avg_dbr_score}{Average debris score, if available.}
\item{genes}{Set of top \code{top_n} marker genes separated by 
 semicolons.}
}
The marker genes are calculated as the genes with the largest 
difference in proportion between the
}
