% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{t_test_sparse}
\alias{t_test_sparse}
\title{Run t test given parameters of the mean, variance, and size.}
\usage{
t_test_sparse(x_mean, x_var, x_n, y_mean, y_var, y_n)
}
\arguments{
\item{x_mean}{Numeric. Mean of group x.}

\item{x_var}{Numeric. Variance of group x.}

\item{x_n}{Numeric. Size of group x.}

\item{y_mean}{Numeric. Mean of group y.}

\item{y_var}{Numeric. Variance of group y.}

\item{y_n}{Numeric. Size of group y.}
}
\value{
list with 2 components:
\describe{
  \item{p}{p-value}
  \item{log2fc}{Log base 2 fold change (x over y)}
}
}
\description{
This t test function rapidly calculates p value and log fold change
given the parameters.
}
